### 2021-09-12 #3721 @whedon
[PRE REVIEW]: GiNaCDE: the high-performance F-expansion and First Integral Methods with C++ library for solving Nonlinear Differential Equations
**Submitting author:** @mithun218 (<a href="http://orcid.org/0000-0002-9678-4625">Mithun Bairagi</a>)
**Repository:** <a href="https://github.com/mithun218/GiNaCDE" target ="_blank">https://github.com/mithun218/GiNaCDE</a>
**Version:** v1.0.0
