+++ JOSS-paper/paper.bib
@@ -5,7 +5,7 @@ volume = {308},
 number = {1},
 pages = {31-36},
 year = {2003},
-doi = {https://doi.org/10.1016/S0375-9601(02)01775-9},
+doi = {10.1016/S0375-9601(02)01775-9},
 url = {https://www.sciencedirect.com/science/article/pii/S0375960102017759},
 author = {Yubin Zhou and Mingliang Wang and Yueming Wang}
 }
@@ -89,7 +89,7 @@ volume = {216},
 number = {10},
 pages = {3064-3072},
 year = {2010},
-doi = {https://doi.org/10.1016/j.amc.2010.04.026},
+doi = {10.1016/j.amc.2010.04.026},
 url = {https://www.sciencedirect.com/science/article/pii/S0096300310004467},
 author = {Zai-yun Zhang and Zhen-hai Liu and Xiu-jin Miao and Yue-zhong Chen}
 }
@@ -101,7 +101,7 @@ volume = {33},
 number = {1},
 pages = {1-12},
 year = {2002},
-doi = {https://doi.org/10.1006/jsco.2001.0494},
+doi = {10.1006/jsco.2001.0494},
 url = {https://www.sciencedirect.com/science/article/pii/S0747717101904948},
 author = {Christian Bauer and Alexander Frink and Richard Kreckel}
 }
@@ -113,7 +113,7 @@ volume = {374},
 number = {2},
 pages = {549-553},
 year = {2011},
-doi = {https://doi.org/10.1016/j.jmaa.2010.08.050},
+doi = {10.1016/j.jmaa.2010.08.050},
 url = {https://www.sciencedirect.com/science/article/pii/S0022247X10007031},
 author = {N. Taghizadeh and M. Mirzazadeh and F. Farahrooz}
 }
@@ -125,7 +125,7 @@ volume = {98},
 number = {3},
 pages = {288-300},
 year = {1996},
-doi = {https://doi.org/10.1016/0010-4655(96)00104-X},
+doi = {10.1016/0010-4655(96)00104-X},
 url = {https://www.sciencedirect.com/science/article/pii/001046559600104X},
 author = {E.J. Parkes and B.R. Duffy}
 }
@@ -137,7 +137,7 @@ volume = {148},
 number = {2},
 pages = {256-266},
 year = {2002},
-doi = {https://doi.org/10.1016/S0010-4655(02)00559-3},
+doi = {10.1016/S0010-4655(02)00559-3},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465502005593},
 author = {Zhi-bin Li and Yin-ping Liu}
 }
@@ -149,7 +149,7 @@ volume = {37},
 number = {6},
 pages = {669-705},
 year = {2004},
-doi = {https://doi.org/10.1016/j.jsc.2003.09.004},
+doi = {10.1016/j.jsc.2003.09.004},
 url = {https://www.sciencedirect.com/science/article/pii/S0747717104000082},
 author = {D. Baldwin and Ü. Göktaş and W. Hereman and L. Hong and R.S. Martino and J.C. Miller}
 }
