### 2021-10-01 #3721 @meg-simula
:wave: Thanks, @mithun218 for the submission. This looks very much ready for review with JOSS.

I'll reach out to reviewers, and get back to you.  Would you have any suggestions for reviewers, for instance from[ this spreadsheet](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822)?
