### 2021-10-01 #3721 @meg-simula
:wave: @peanutfun @balos1 @ketch would any of you be willing to review this submission for JOSS?
