+++ src/simplify.cpp
@@ -311,7 +311,7 @@ ex expandinv::operator()(const ex& e)
 /**doing factors of fractional power bases, function arguments. **/
 ex arguSimplify::operator()(const ex &e)
 {
-    return (Simplify(Factor(e, factor_all))).map(*this);
+    return (Simplify(Factor(e))).map(*this);
 }
 
 /** doing number simplify. **/
@@ -394,25 +394,25 @@ ex powBaseSubsLessThanDegLvl_1::operator()(const ex& _e)
 /** It is level 1 implementation of fracPowBasSubs. **/
 ex fracPowBasSubsLvl_1::operator()(const ex& e)
 {
-    if(is_a<power>(e)&&Denom(e.op(1))!=_ex1)
+    if(is_a<power>(e)&&denom(e.op(1))!=_ex1)
     {
         numer_denomClt = (e.op(0)).numer_denom();
 
         tem = Simplify(collect_common_factors(Factor(Simplify(expand(numer_denomClt.op(0)))))/
                        collect_common_factors(Factor(Simplify(expand(numer_denomClt.op(1))))));
 
-        if((!baseCltLvl_1.empty() && baseCltLvl_1.find(pow(tem,pow(Denom(e.op(1)),_ex_1)))==baseCltLvl_1.end())
+        if((!baseCltLvl_1.empty() && baseCltLvl_1.find(pow(tem,pow(denom(e.op(1)),_ex_1)))==baseCltLvl_1.end())
             || baseCltLvl_1.empty())
         {
             j=j+1;
 
             str = "genSymb1_" + to_string(j);
             expr = reader(str);
-            baseCltLvl_1[pow(tem,pow(Denom(e.op(1)),_ex_1))]=expr;
+            baseCltLvl_1[pow(tem,pow(denom(e.op(1)),_ex_1))]=expr;
         }
 
-        if(!baseCltLvl_1.empty() && baseCltLvl_1.find(pow(tem,pow(Denom(e.op(1)),_ex_1)))!=baseCltLvl_1.end())
-            return pow(baseCltLvl_1[pow(tem,pow(Denom(e.op(1)),_ex_1))],Numer(e.op(1)));
+        if(!baseCltLvl_1.empty() && baseCltLvl_1.find(pow(tem,pow(denom(e.op(1)),_ex_1)))!=baseCltLvl_1.end())
+            return pow(baseCltLvl_1[pow(tem,pow(denom(e.op(1)),_ex_1))],numer(e.op(1)));
     }
 
     return e.map(*this);
@@ -423,7 +423,7 @@ ex fracPowBasSubsLvl_1::operator()(const ex& e)
  * such as numeric, symbols, add container. This class has been used in Factor function before factoring.    **/
 ex fracPowBasSubsFactor::operator()(const ex& e)
 {
-    if(is_a<power>(e)&&Denom(e.op(1))!=_ex1)
+    if(is_a<power>(e)&&denom(e.op(1))!=_ex1)
     {
 
         numer_denomClt = (e.op(0)).numer_denom();
@@ -431,18 +431,18 @@ ex fracPowBasSubsFactor::operator()(const ex& e)
         tem = Simplify(collect_common_factors(Factor(Simplify(expand(numer_denomClt.op(0)))))/
                        collect_common_factors(Factor(Simplify(expand(numer_denomClt.op(1))))));
 
-        if((!baseClt.empty() && baseClt.find(pow(tem,pow(Denom(e.op(1)),_ex_1)))==baseClt.end())
+        if((!baseClt.empty() && baseClt.find(pow(tem,pow(denom(e.op(1)),_ex_1)))==baseClt.end())
             || baseClt.empty())
         {
             j=j+1;
 
[…]
