### 2021-11-03 #3721 @meg-simula
:wave: @justinsgray @sivaramambikasaran @crstnbr @schruste would any of you be willing to review this submission for JOSS?
