### 2021-11-03 #3885 @whedon
[REVIEW]: GiNaCDE: the high-performance F-expansion and First Integral Methods with C++ library for solving Nonlinear Differential Equations
**Submitting author:** @mithun218 (<a href="http://orcid.org/0000-0002-9678-4625">Mithun Bairagi</a>)
**Repository:** https://github.com/mithun218/GiNaCDE
**Branch with paper.md** (empty if default branch):
**Version:** v1.5.0
