### 2021-11-08 mithun218/GiNaCDE#1 @peanutfun
Clarify licensee in LICENSE.txt
### Description

The [`LICENSE.txt`](https://github.com/mithun218/GiNaCDE/blob/f332d194ec753ab2afa7e14c983e2d3f0699bfbf/LICENSE.txt) file lists the licensee "Barak Shoshany", although the author of this repository appears to be @mithun218 ("Mithun Bairagi" according to the JOSS submission). I guess this is a simple copy-paste error? Please clarify and update the license file accordingly.

### Related issues

openjournals/joss-reviews#3885
