### 2021-11-08 mithun218/GiNaCDE#2 @peanutfun
Make `desolve` return a function expression
### Description

The main feature of GiNaCDE is the [`desolve`](https://github.com/mithun218/GiNaCDE/blob/f332d194ec753ab2afa7e14c983e2d3f0699bfbf/src/desolve.h#L43) function which solves a given differential equation. However, this function does not return the resulting GiNaC expression `ex`, but only an integer status. The expression is written to a file and discarded. This severely hampers the usability of the GiNaCDE library, as the result of `desolve` cannot be used within a program for further computations (e.g., evaluating it, integrating it over an interval, ...).

Is there maybe a technical reason why the result is not also returned by the function?

### Proposal

Make `desolve` return the solved expression, or throw an exception if there was an error.

### Related issues

openjournals/joss-reviews#3885
