### 2021-11-08 mithun218/GiNaCDE#3 @peanutfun
Make tests actually check the computation results
### Description

The [tests](https://github.com/mithun218/GiNaCDE/tree/master/test) in this repository contain source files that solve certain differential equations and write the results into a file. However, the results themselves are not checked. As such, the tests only check if the differential equations can be solved without throwing an error, but this does not verify the the results are *correct*.

As stated in the `README.md`, the GiNaC library might not return the exact same output of an expression due to some unpredictable internal ordering of symbols. So it is unsafe to check the output file contents. However, one could evaluate the resulting expressions within the code. This would either require that the expressions are returned by `desolve` (see #2) or that a file reader is added, which reads the output files of GiNaCDE and returns a GiNaC expression that can be evaluated.

### Proposal

The tests should evaluate and this verify the results of `desolve` at least for few known values. This requires a way of retrieving the result as expression `ex`.

In the documentation of GiNaC I did not find a way to compare two expressions `ex`. However, one can numerically evaluate them with [`ex::evalf`](https://www.ginac.de/reference/classGiNaC_1_1ex.html#a6ecfce5165142ab928503dfce220ed2e). If one knows the correct result of a differential equation, one could check for specific values like its roots.

### Related issues

* `desolve` result is not returned: #2
* JOSS review: openjournals/joss-reviews#3885
