### 2021-11-08 mithun218/GiNaCDE#4 @peanutfun
Consider non-proprietary output format
### Description

GiNaCDE writes output in the Maple and Mathematica language syntax. Both are proprietary. Although https://www.wolframalpha.com technically understands Mathematica, its use in this case is limited, as the expressions found by GiNaCDE are usually to complicated to be tackled by the online tool. I think GiNaCDE would attract a larger audience by supporting an additional output format that does not require a software subscription.

### Proposal

Add a non-proprietary output format to GiNaCDE. This can serve two purposes: Better human readability and more use cases in follow-up computations. Two things come to my mind based on the [print contexts](https://www.ginac.de/reference/classGiNaC_1_1print__context.html) implemented in GiNaC:

* Python: Write output that can be parsed by a Python interpreter.
* LaTeX: Write output that can be compiled into a PDF for better readability.

### Related issues

openjournals/joss-reviews#3885
