+++ examples/Eckhaus_FIM.txt
@@ -1,43 +1,40 @@
                               =====First Integral Method======                         
-Equations are written in MAPLE language.
+Equations are written in MATHEMATICA language.
 ----------------------------------------------------------------------------------------------------
 
-Input equation is: conjugate(u(t,x))^2*u(t,x)^3+2*u(t,x)*diff(conjugate(u(t,x))*u(t,x),x$1)+I*diff(u(t,x),t$1)+diff(u(t,x),x$2) = 0;
-Real part of Diff. Equ.: U(xi)^5+k_1^2*diff(U(xi),xi$2)-U(xi)*p_0+4*U(xi)^2*k_1*diff(U(xi),xi$1)-U(xi)*p_1^2 = 0;
-Imaginary part of Diff. Equ.: k_0*diff(U(xi),xi$1)+2*k_1*diff(U(xi),xi$1)*p_1 = 0;
-Imaginary part of Diff. Equ. is integrable.
-After integration, Imaginary part: 2*U(xi)*k_1*p_1+k_0*U(xi) = 0;
-We derive solutions of real part of Diff. Equ. with the condition:
- k_0 = -2*k_1*p_1;
-u = U(xi)*exp(I*(t*p_0+x*p_1)),
-where xi = k_1*x-2*k_1*t*p_1;
+Input equation is: I*D[u[t,x],{t,1}]+2*u[t,x]*D[u[t,x]*Conjugate[u[t,x]],{x,1}]+D[u[t,x],{x,2}]+u[t,x]^3*Conjugate[u[t,x]]^2 = 0;
+Real part of Diff. Equ.: -U[xi]*b+U[xi]^5+k^2*D[U[xi],{xi,2}]-U[xi]*a^2+4*k*D[U[xi],{xi,1}]*U[xi]^2 = 0;
+Imaginary part of Diff. Equ.: 0 = 0;
+The Diff. Equ. becomes: -U[xi]*b+U[xi]^5+k^2*D[U[xi],{xi,2}]-U[xi]*a^2+4*k*D[U[xi],{xi,1}]*U[xi]^2 = 0;
+u = U[xi]*Exp[I*(a*x+t*b)],
+where xi = k*x-2*k*t*a;
 The value of N is: 1;
 
 ****************************************************************************************************
 
--a_1*X_^5+a_1*X_*p_0+a_1*X_*p_1^2+Y_*Diff(a_0,X_,1)*k_1^2+Y_^2*k_1^2*Diff(a_1,X_,1) = k_1^2*g_*a_0+Y_^2*a_1*k_1^2*h_+Y_*a_1*k_1^2*g_+Y_*k_1^2*a_0*h_+4*Y_*a_1*X_^2*k_1
[…]
