+++ examples/eckhaus_mF.txt
@@ -2,68 +2,59 @@
 Equations are written in MAPLE language.
 ----------------------------------------------------------------------------------------------------
 
-Input equation is: I*diff(u(t,x),t$1)+2*u(t,x)*diff(u(t,x)*conjugate(u(t,x)),x$1)+diff(u(t,x),x$2)+u(t,x)^3*conjugate(u(t,x))^2 = 0;
-Real part of Diff. Equ.: -U(xi)*b+U(xi)^5+k^2*diff(U(xi),xi$2)-U(xi)*a^2+4*k*diff(U(xi),xi$1)*U(xi)^2 = 0;
-Imaginary part of Diff. Equ.: 0 = 0;
-The Diff. Equ. becomes: -U(xi)*b+U(xi)^5+k^2*diff(U(xi),xi$2)-U(xi)*a^2+4*k*diff(U(xi),xi$1)*U(xi)^2 = 0;
-u = U(xi)*exp(I*(a*x+t*b)),
-where xi = k*x-2*k*t*a;
+Input equation is: conjugate(u(t,x))^2*u(t,x)^3+2*u(t,x)*diff(conjugate(u(t,x))*u(t,x),x$1)+I*diff(u(t,x),t$1)+diff(u(t,x),x$2) = 0;
+Real part of Diff. Equ.: U(xi)^5+k_1^2*diff(U(xi),xi$2)-U(xi)*p_0+4*U(xi)^2*k_1*diff(U(xi),xi$1)-U(xi)*p_1^2 = 0;
+Imaginary part of Diff. Equ.: k_0*diff(U(xi),xi$1)+2*k_1*diff(U(xi),xi$1)*p_1 = 0;
+Imaginary part of Diff. Equ. is integrable.
+After integration, Imaginary part: 2*U(xi)*k_1*p_1+k_0*U(xi) = 0;
+We derive solutions of real part of Diff. Equ. with the condition:
+ k_0 = -2*k_1*p_1;
+u = U(xi)*exp(I*(t*p_0+x*p_1)),
+where xi = k_1*x-2*k_1*t*p_1;
 The value of N is: 1/2;
-U = a_0+b_1*F^(-1/2)+a_1*sqrt(F);
-The first-order nonlinear ODE: diff(F(xi),xi) = A_0+A_2*F^2;
+U = b_1*F^(-1/2)+a_1*sqrt(F)+a_0;
+The first-order nonlinear ODE: diff(F(xi),xi) = F^2*A_2+F*A_1;
 
 ****************************************************************************************************
 
 The system of algebraic equations are: 
-F^(0/2): 4*b_1^5+3*A_0^2*k^2*b_1-8*A_0*k*b_1^3 = 0;
-F^(1/2): 20*a_0*b_1^4-16*A_0*a_0*k*b_1^2 = 0;
[…]
