### 2021-11-08 mithun218/GiNaCDE#5 @peanutfun
Add OS-dependent installation instructions
### Description

This repository currently provides tarballs for the required dependencies CLN and GiNaC, and users are instructed to install both before installing GiNaCDE. The optional dependency GTK3 is advertised to be downloaded via APT. In total, the given instructions only work on Ubuntu or Debian. However, on both of these systems, all dependencies are available via APT, and including the tarballs would not be required. Additionally, all dependencies can also be retrieved on macOS via the most common package managers [Homebrew](https://brew.sh/) and [MacPorts](https://ports.macports.org/). On Windows, on the other hand, the current installation instructions do not apply.

### Proposal

Simplify installation instructions by relying on package managers as far as possible. For example, CLN, GiNaC and GTK3 are installed via APT through
```
apt install libcln-dev libginac-dev libgtk-3-dev
```

and on Homebrew via
```
brew install cln ginac gtk+3
```

Finally, separate instructions would need to be added for Windows.

### Related issues

* #6
* openjournals/joss-reviews#3885
