### 2021-11-08 mithun218/GiNaCDE#6 @peanutfun
Remove tarballs from repository
### Description

This repository currently contains tarballs of the main software dependencies and a pre-compiled GUI. Tarballs are binary code that cannot be interpreted by Git. This leads to a bloat in repository size, especially when multiple versions of the tarballs are committed. In my perception, including the tarballs is not necessary: The dependencies can be installed through usual package managers (see #5), or downloaded from their respective repositories or websites.

The GUI, on the other hand, is pre-compiled for Windows and cannot be used for other platforms. This goes against the idea of a repository containing platform-independent source code. While I very much like the idea of ready-to-use binaries, I recommend uploading the GUI package somewhere else, e.g. directly to https://zenodo.org/ or simply to a Google Drive.

### Proposal

* Remove tarballs of dependency packages, updating installation instructions accordingly (in line with #5).
* Remove GUI package tarball. Upload it to a suitable different repository, and provide a link in the `README.md`.

### Related issues

* Update installation instructions: #5
* openjournals/joss-reviews#3885
