### 2021-11-08 mithun218/GiNaCDE#7 @peanutfun
Add missing LaTeX source files
### Description

This repository contains several PDF files without respective LaTeX source file counterparts. Adding them to the repository would increase transparency and simplify future maintenance. Additionally (and optionally), CMake provides a [`FindLATEX.cmake`](https://cmake.org/cmake/help/latest/module/FindLATEX.html) module which can be used to re-compile the PDFs through the build system.

### Proposal

* Add missing LaTeX source files for PDFs included in this repository.
* *optional:* Let CMake search for LaTeX and add optional target `make_latex_docs` (or similar), that compiles the LaTeX files.

### Related issues

openjournals/joss-reviews#3885
