### 2021-11-08 #3885 @peanutfun
@mithun218 Before I can summarize I need a bit of clarification on how to use and interpret the output of GiNaCDE. I very much like the extensive documentation you provide, with an in-depth look at the mathematical methods and clear instructions on how to use GiNaCDE. However, it does not explain how to use the resulting output (files). That's why I have some trouble verifying if the software computes the correct solutions.

Admittedly, I am not too familiar with the methods you use and the problems you try to solve, so I might have a wrong impression or perspective. Considering the [example usage](https://github.com/mithun218/GiNaCDE#example-usage) you propose in the `README.md`, how would a typical workflow look like after calculating a solution? In particular: What to do with the resulting output `.txt` file? Its contents do not seem self-explanatory to me.

GiNaCDE provides the option to output either in the Maple or the Mathematica format. However, the output files also contain plain text and cannot simply be "imported" into either of the programs (Please, correct me if I'm wrong!). How could I use the output of GiNaCDE in them? Would I need to copy-paste the expressions, leaving out the text? How do I get a "computational" representation of the results to continue on with?

I myself own neither of the programs. How can I make use of the results in my case?
