+++ examples/eckhaus_mF.txt
@@ -2,59 +2,68 @@
 Equations are written in MAPLE language.
 ----------------------------------------------------------------------------------------------------
 
-Input equation is: conjugate(u(t,x))^2*u(t,x)^3+2*u(t,x)*diff(conjugate(u(t,x))*u(t,x),x$1)+I*diff(u(t,x),t$1)+diff(u(t,x),x$2) = 0;
-Real part of Diff. Equ.: U(xi)^5+k_1^2*diff(U(xi),xi$2)-U(xi)*p_0+4*U(xi)^2*k_1*diff(U(xi),xi$1)-U(xi)*p_1^2 = 0;
-Imaginary part of Diff. Equ.: k_0*diff(U(xi),xi$1)+2*k_1*diff(U(xi),xi$1)*p_1 = 0;
-Imaginary part of Diff. Equ. is integrable.
-After integration, Imaginary part: 2*U(xi)*k_1*p_1+k_0*U(xi) = 0;
-We derive solutions of real part of Diff. Equ. with the condition:
- k_0 = -2*k_1*p_1;
-u = U(xi)*exp(I*(t*p_0+x*p_1)),
-where xi = k_1*x-2*k_1*t*p_1;
+Input equation is: I*diff(u(x,t),t$1)+diff(u(x,t),x$2)+2*u(x,t)*diff(u(x,t)*conjugate(u(x,t)),x$1)+u(x,t)^3*conjugate(u(x,t))^2 = 0;
+Real part of Diff. Equ.: -a^2*U(xi)+4*k*U(xi)^2*diff(U(xi),xi$1)+U(xi)^5-b*U(xi)+k^2*diff(U(xi),xi$2) = 0;
+Imaginary part of Diff. Equ.: 0 = 0;
+The Diff. Equ. becomes: -a^2*U(xi)+4*k*U(xi)^2*diff(U(xi),xi$1)+U(xi)^5-b*U(xi)+k^2*diff(U(xi),xi$2) = 0;
+u = U(xi)*exp(I*(b*t+a*x)),
+where xi = k*x-2*a*k*t;
 The value of N is: 1/2;
-U = b_1*F^(-1/2)+a_1*sqrt(F)+a_0;
-The first-order nonlinear ODE: diff(F(xi),xi) = F^2*A_2+F*A_1;
+U = a_1*sqrt(F)+b_1*F^(-1/2)+a_0;
+The first-order nonlinear ODE: diff(F(xi),xi) = A_2*F^2+A_0;
 
 ****************************************************************************************************
 
 The system of algebraic equations are: 
-F^(0/2): 4*b_1^5 = 0;
-F^(1/2): 20*b_1^4*a_0 = 0;
-F^(2/2): 40*b_1^3*a_0^2+20*a_1*b_1^4-8*k_1*b_1^3*A_1 = 0;
[…]
