### 2021-11-09 #3885 @mithun218
@peanutfun thanks so much for your review and comments.
I am happy to edit the paper to incorporate your comments.

The GiNaCDE software determines exact solutions with calculating steps, and the results are saved in an output .txt file in the Maple or the Mathematica format. In order to use the output exact solutions in Maple or Mathematica software, we have to copy the exact solutions with the conditions of solutions (which are also provided in the output .txt file) from the output .txt file. Then, we can verify the exact solutions using Maple or Mathematica software. To verify the exact solutions at first, we have to substitute the exact solutions with the corresponding conditions of the solutions into the input NLPDE and then simplify it. After simplification, if it is equal to zero, then the solution is correct. Otherwise, the solution is not valid.

Besides the output .txt file, the solutions of the NLPDE and the corresponding conditions are also collected by the variables _solutionClt_ and _constraints_. We have explained about these variables in the section **11. INPUT AND OUTPUT OF LIBRARY** of [the documentation file](https://github.com/mithun218/GiNaCDE/blob/master/doc/documentation.pdf). Using these variables, one can easily locate and copy the solutions with the conditions of the solutions in the Maple or the Mathematica software.

 Unfortunately, currently, GiNaCDE is unable to verify the solutions without Maple or the Mathematica software. But this will be a very interesting feature of the GiNaCDE if the GiNaCDE itself is able to verify the solutions. Now I am planning that I shall add this feature in the next release of GiNaCDE but this will take some time.

I will try to resolve all the above issues (mentioned by you) in the next few days.
