### 2021-11-09 mithun218/GiNaCDE#1 @mithun218
You are right. It was a simple copy-paste error. The licensee will be Mithun Bairagi (the author of this repository). I have fixed this issue in [2594454](https://github.com/mithun218/GiNaCDE/commit/259445470b79bc4b65260b34a6c23c58088e1e42).
