### 2021-11-10 fd2d7b3
Add examples to CMake configuration
* Register examples as executables that are `EXCLUDED_FROM_ALL`
  (not built by default).
* Add meta-target `examples` for building all examples.
* Fix includes for all examples.
