+++ examples/example1.cpp
@@ -8,7 +8,7 @@
 
 
 
-#include <GiNaCDE/GiNaCDE.h>
+#include <GiNaCDE.h>
 
 
 int main()
