### 2021-11-10 efde896
Update README.md
Add instructions for building examples.
