### 2021-11-10 5623a23
Fix formatting issue in KS.cpp
