### 2021-11-10 mithun218/GiNaCDE#10 @peanutfun
Add doxygen documentation to header files
### Description

[Doxygen](https://www.doxygen.nl/index.html) is a common and very useful tool for building a documentation of C++ source code based on reading [code comments in a specific format](https://www.doxygen.nl/manual/docblocks.html). It helps keeping the code documentation "with the code", meaning that one is less likely to update the code, but not its documentation.

GiNaCDE provides a type of "public API" by exposing certain variables and functions through its header files. Most of them are currently listed in the [`documentation.pdf`](https://github.com/mithun218/GiNaCDE/blob/master/doc/documentation.pdf) under "Some useful commands in library". I suggest documenting at least these variables and commands with Doxygen as well.

CMake has a [module](https://cmake.org/cmake/help/latest/module/FindDoxygen.html) which searches for Doxygen and provides useful functions for building the documentation.

### Proposal

* Add doxygen docstrings to variables and commands in the GiNaCDE headers that are considered "public API", i.e. interacted with by users.
* Make CMake build the Doxygen documentation (if Doxygen is found on the host system)

### Related issues

openjournals/joss-reviews#3885
