### 2021-11-10 mithun218/GiNaCDE#12 @peanutfun
Improve the JOSS paper
### Summary

The JOSS paper is clear and does a good job in giving a statement of need for the software. Nonetheless, I noticed some points that could be improved, mostly to make it more concise.

### Suggestions

* I have noticed some grammatical and spelling errors in the paper. Since English is not my mother tongue, I suggest you use a spell and grammar checker on the paper.  You could use https://www.grammarly.com/, MS Word, or some editor plugin of choice.
* The "Algorithms" section seems redundant to me. You mention all methods and references in previous sections already. Your explanation of the formation of the methods from this section would fit well into the first paragraph of the "Statement of Need" section.
* The references for F-expansion (`[@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn]`) and first integral methods (`[@fim0][@mirza][@complexTwt1]`) appear several times throughout the paper and even several times in single sections. Please avoid repeated citations of the same paper and cite a source only once per section. Also, please clarify which references explain the *modified* F-expansion, and which the "regular" F-expansion method.
* Avoid "etc." after listings and use "such as" or "e.g." before listings instead.
* > GiNaCDE is a free and open-source C++ library, which is an NLPDE (Nonlinear Partial Differential Equation) and NLODE (Nonlinear Ordinary Differential Equation) solver based on [...]

   Avoid "which is" here, as GiNaCDE being a solver is more important than being a C++ library. Suggestion:

   > GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on [...]

    (or similar)
* > On the other hand, F-expansion, mF-expansion, and FIM are different kinds of methods, which can overcome the limitations of the function-expansion method.

    Please add a suitable reference for this statement.
* > We have successfully solved more than 30 NLPDEs, such as Eckhaus equation, [...]

   This paragraph would fit better into the summary. Instead of giving a number of equations, I suggest listing (only) the equations given in the examples in the repository. Also, please state the unabbreviated names of these equations instead of acronyms.
* > [...] and it gives results not more than 100 seconds.

   It is unclear here if this time refers to solving one or all equations you mention, and if compile times are included. Further, executable run times are highly dependent on the system hardware. Please remove this claim altogether or make it more precise, e.g. by benchmarking with similar methods from other packages.

### Related issues

openjournals/joss-reviews#3885
