### 2021-11-10 #3885 @peanutfun
@mithun218 thank you for your quick reply. Indeed, I missed the part on `solutionClt` and `constraints`. Let me summarize my review and get back to my question:

You obviously put a lot of effort into implementing the F-expansion and first integral methods in an open-source framework. Your documentation focuses strongly on the intricate mathematics involved in these methods, and admittedly, they are a bit out of my reach (I'm a physicist 🙈 ). Your paper makes a good point why GiNaCDE is needed, especially in the open-source context. And I really like the GUI, it works intuitively and is a great option for users that are less familiar with the command line and C++ programming.

While GiNaCDE probably serves you well in your own workflow, the focus shifts from functionality to usability when it comes to a software publication. The goal is to address a wider audience for the software, and even people who are unfamiliar with the methods (like me) need to quickly understand how the software is installed, how they interact with it, how they use its results in their own workflow, and that everything works as intended. I find that GiNaCDE as a software package lacks this wider perspective to some degree.

Most importantly, it is very difficult for me (as somebody unfamiliar with the mathematics) to verify the results of GiNaCDE. This verification is best done automatically in tests, but as I pointed out in mithun218/GiNaCDE#3, the currently implemented tests do not verify the results. In your reply you point out that Maple or Mathematica are required for this. But you also state that

> To verify the exact solutions at first, we have to substitute the exact solutions with the corresponding conditions of the solutions into the input NLPDE and then simplify it. After simplification, if it is equal to zero, then the solution is correct. Otherwise, the solution is not valid.

Isn't this procedure also possible within GiNaCDE? Could this be used to test your results in a general way? If not, you could have the tests compute solutions for relatively simple ODEs or PDEs, and then evaluate the resulting expressions for some known function values like roots, maxima, etc. I focus on this because (automated) tests for verifying the software are part of the JOSS review criteria, and I cannot check this off until there is at least a manual procedure documented on how to verify the results of the library.

Also with regard to usability, I think the installation instructions need to be improved to account for different operating systems, and the usage instructions should show how other CMake projects can incorporate GiNaCDE. This leads me to the following list of issues we need to resolve to match the JOSS review criteria from my perspective:

### Issues for meeting review criteria

- [ ] Improve the automated tests to verify computation results (or add manual instructions on how to do so): mithun218/GiNaCDE#3
- [ ] Simplify installation instructions for Linux, and possibly add instructions for macOS and Windows: mithun218/GiNaCDE#5
- [ ] Add instructions on how to use GiNaCDE in downstream modules: mithun218/GiNaCDE#8
- [ ] Incorporate or conclusively reject suggestions to the JOSS paper: mithun218/GiNaCDE#12

Personally, I consider all other issues I opened in the GiNaCDE repository *non-crucial* for accepting your submission.
