### 2021-11-12 69b82a2
remove the output of the examples (joss-reviews #3885)
