+++ examples/cahnAllen_mF.txt
@@ -1,162 +1,0 @@
-                              =====Modified F-Expansion Method======                         
-Equations are written in MATHEMATICA language.
-----------------------------------------------------------------------------------------------------
-
-Input equation is: D[u[t,x],{t,1}]-u[t,x]+u[t,x]^3-D[u[t,x],{x,2}] = 0;
-The Diff. Equ. becomes: D[U[xi],{xi,1}]*k_0-U[xi]-k_1^2*D[U[xi],{xi,2}]+U[xi]^3 = 0;
-u = U[xi], 
-where xi = t*k_0+k_1*x;
-The value of N is: 1;
-U = a_0+a_1*F;
-The first-order nonlinear ODE: D[F[xi],xi] = A_1*F+A_2*F^2;
-
-****************************************************************************************************
-
-The system of algebraic equations are: 
-F^0: -a_0+a_0^3 = 0;
-F^1: A_1*a_1*k_0-k_1^2*A_1^2*a_1-a_1+3*a_0^2*a_1 = 0;
-F^2: -3*k_1^2*A_1*a_1*A_2+3*a_0*a_1^2+a_1*A_2*k_0 = 0;
-F^3: -2*k_1^2*a_1*A_2^2+a_1^3 = 0;
- In the following results C_ is an arbitrary constant.
-
-****************************************************************************************************
-
-solving above system of equations for variables {k_0,k_1,a_0,a_1}->
-
-{k_1=1/2*A_1^(-1)*Sqrt[2],k_0=-3/2*A_1^(-1),a_1=-A_1^(-1)*A_2,a_0=-1}
-U = -1-A_1^(-1)*A_2*F,
-where F is the solution of
-D[F[xi],xi] = A_1*F+A_2*F^2;
-Solution[s] of input Diff. Equ. is (are)=>
-u = -1+(-1+Cosh[C_*A_1+1/2*A_1*(A_1^(-1)*Sqrt[2]*x-3*A_1^(-1)*t)]*A_2+Sinh[C_*A_1+1/2*A_1*(A_1^(-1)*Sqrt[2]*x-3*A_1^(-1)*t)]*A_2)^(-1)*A_2*(Cosh[C_*A_1+1/2*A_1*(A_1^(-1)*Sqrt[2]*x-3*A_1^(-1)*t)]+Sinh[C_*A_1+1/2*A_1*(A_1^(-1)*Sqrt[2]*x-3*A_1^(-1)*t)]);
-u = -1+(A_1^(-1)*A_2-Exp[-1/2*A_1*(A_1^(-1)*Sqrt[2]*x-3*A_1^(-1)*t)]*C_)^(-1)*A_1^(-1)*A_2;
[…]
