+++ examples/7thorder_FIM.txt
@@ -1,16 +1,0 @@
-                              =====First Integral Method======                         
-Equations are written in MAPLE language.
-----------------------------------------------------------------------------------------------------
-
-Input equation is: diff(126*diff(u(t,x),x$2)*u(t,x)^2+63*u(t,x)^4+21*diff(u(t,x),x$4)*u(t,x)+63*diff(u(t,x),x$1)^2*u(t,x)+21*diff(u(t,x),x$1)*diff(u(t,x),x$3)+21*diff(u(t,x),x$2)^2+diff(u(t,x),x$6),x$1)+diff(u(t,x),t$1) = 0;
-The Diff. Equ. is integrable;
-The assigned value(s) to the integrating constant(s)-> 
-ic_1: 0
-The Diff. Equ. becomes: diff(U(xi),xi$6)*k_1^7+21*diff(U(xi),xi$2)^2*k_1^5+k_0*U(xi)+63*U(xi)^4*k_1+126*U(xi)^2*diff(U(xi),xi$2)*k_1^3+21*diff(U(xi),xi$1)*k_1^5*diff(U(xi),xi$3)+63*diff(U(xi),xi$1)^2*U(xi)*k_1^3+21*diff(U(xi),xi$4)*U(xi)*k_1^5 = 0;
-u = U(xi), 
-where xi = x*k_1+t*k_0;
-The value of N is: 1;
-
-****************************************************************************************************
-
-Evaluation stop: order of ode should be 2;
