+++ examples/breakingSoliton_mF.txt
@@ -1,68 +1,0 @@
-                              =====Modified F-Expansion Method======                         
-Equations are written in MAPLE language.
-----------------------------------------------------------------------------------------------------
-
-Input equation is: -2*diff(u(y,t,x),y$1)*diff(u(y,t,x),x$2)+diff(diff(u(y,t,x),x$3),y$1)-4*diff(diff(u(y,t,x),x$1),y$1)*diff(u(y,t,x),x$1)+diff(diff(u(y,t,x),x$1),t$1) = 0;
-The Diff. Equ. becomes: k_1^3*k_2*diff(U(xi),xi$4)+k_1*diff(U(xi),xi$2)*k_0-6*k_1^2*diff(U(xi),xi$2)*diff(U(xi),xi$1)*k_2 = 0;
-u = U(xi), 
-where xi = y*k_2+t*k_0+k_1*x;
-The value of N is: 1;
-U = a_0+b_1*F^(-1)+a_1*F;
-The first-order nonlinear ODE: diff(F(xi),xi) = A_1*F+A_2*F^2;
-
-****************************************************************************************************
-
-The system of algebraic equations are: 
-F^0: 6*k_1^2*A_1^3*k_2*b_1^2 = 0;
-F^1: k_1^3*A_1^4*k_2*b_1+12*k_1^2*A_1^2*k_2*A_2*b_1^2+k_1*A_1^2*b_1*k_0 = 0;
-F^2: 6*k_1^2*A_1*k_2*A_2^2*b_1^2+k_1*A_1*A_2*b_1*k_0+k_1^3*A_1^3*k_2*A_2*b_1 = 0;
-F^3: k_1^3*A_1^4*a_1*k_2+k_1*A_1^2*a_1*k_0+12*k_1^2*A_1^2*a_1*k_2*A_2*b_1 = 0;
-F^4: -6*k_1^2*A_1^3*a_1^2*k_2+24*k_1^2*A_1*a_1*k_2*A_2^2*b_1+15*k_1^3*A_1^3*a_1*k_2*A_2+3*k_1*A_1*a_1*A_2*k_0 = 0;
-F^5: 2*k_1*a_1*A_2^2*k_0-24*k_1^2*A_1^2*a_1^2*k_2*A_2+50*k_1^3*A_1^2*a_1*k_2*A_2^2+12*k_1^2*a_1*k_2*A_2^3*b_1 = 0;
-F^6: 60*k_1^3*A_1*a_1*k_2*A_2^3-30*k_1^2*A_1*a_1^2*k_2*A_2^2 = 0;
-F^7: -12*k_1^2*a_1^2*k_2*A_2^3+24*k_1^3*a_1*k_2*A_2^4 = 0;
- In the following results C_ is an arbitrary constant.
-
-****************************************************************************************************
-
-solving above system of equations for variables {k_0,k_1,k_2,a_0,a_1,b_1}->
-
-{k_1=0}
-U = a_0+b_1*F^(-1)+a_1*F,
-where F is the solution of
-diff(F(xi),xi) = A_1*F+A_2*F^2;
-solution(s) of input Diff. Equ. is (are)=>
[…]
