+++ examples/KDV_Fex.txt
@@ -1,128 +1,0 @@
-                              =====F-Expansion Method======                         
-Equations are written in MATHEMATICA language.
-----------------------------------------------------------------------------------------------------
-
-Input equation is: D[u[t,x],{t,1}]-u^2*a*D[u[t,x],{x,1}]+D[u[t,x],{x,3}] = 0;
-The Diff. Equ. is integrable;
-The assigned Value[s] to the integrating Constant[s]-> 
-ic_1: 0
-The Diff. Equ. becomes: k_1^3*D[U[xi],{xi,2}]+U[xi]*k_0-1/3*k_1*U[xi]^3*a = 0;
-u = U[xi], 
-where xi = t*k_0+k_1*x;
-The value of N is: 1;
-U = a_0+b_1*F^(-1)+a_1*F;
-The first-order nonlinear ODE: D[F[xi],xi] = Sqrt[A_0+A_4*F^4+A_2*F^2];
-
-****************************************************************************************************
-
-The system of algebraic equations are: 
-Diff[F,xi,1]^0F^0: -k_1*a*b_1^3+6*A_0*k_1^3*b_1 = 0;
-Diff[F,xi,1]^0F^1: -3*a_0*k_1*a*b_1^2 = 0;
-Diff[F,xi,1]^0F^2: -3*a_0^2*k_1*a*b_1+3*k_1^3*A_2*b_1-3*k_1*a_1*a*b_1^2+3*b_1*k_0 = 0;
-Diff[F,xi,1]^0F^3: 3*a_0*k_0-a_0^3*k_1*a-6*a_0*k_1*a_1*a*b_1 = 0;
-Diff[F,xi,1]^0F^4: 3*a_1*k_0+3*k_1^3*a_1*A_2-3*k_1*a_1^2*a*b_1-3*a_0^2*k_1*a_1*a = 0;
-Diff[F,xi,1]^0F^5: -3*a_0*k_1*a_1^2*a = 0;
-Diff[F,xi,1]^0F^6: -k_1*a_1^3*a+6*k_1^3*A_4*a_1 = 0;
- In the following results C_ is an arbitrary constant.
-
-****************************************************************************************************
-
-solving above system of equations for variables {k_0,k_1,a,a_0,a_1,b_1}->
-
[…]

[…]
