### 2021-11-14 931c91e
improve simplify (joss-reviews #3885)
