+++ src/F_expns_methd.cpp
@@ -262,6 +262,7 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
     int odetype;
     ex tra_wave_coordSubs,tw_coordiPhaseSubs = _ex0;
 
+    int solNum = 1; // counts solution numbers
 
     for(set<lst, ex_is_less>::const_iterator it = solu_set_clt.begin(); it != solu_set_clt.end(); it++)
     {
@@ -398,14 +399,15 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
                                 {
                                     if(!is_a<lst>(*it1))
                                     {
-                                        solutions << dpndt_var << " = "  << solu_form_F_subs <<";"<< endl;
+                                        solutions<<"solution #"<<solNum<<"  " << dpndt_var << " = "  << solu_form_F_subs <<";"<< endl;
+                                        solNum = solNum + 1;
 
                                         solutionClt[solutionClt.size()-1].append(dpndt_var  ==  solu_form_F_subs);
                                     }
                                     else // for handling solutions with conditions
                                     {
                                         lst temList = {};
-                                        solutions << dpndt_var << " = "  << solu_form_F_subs <<" (with condition(s) ";
+                                        solutions<<"solution #"<<solNum<<"  " << dpndt_var << " = "  << solu_form_F_subs <<" (with condition(s) ";
                                         temList.append(dpndt_var  ==  solu_form_F_subs);
 
                                         for(size_t i=1; i<nops(*it1); i++)
@@ -414,6 +416,8 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
                                             temList.append((*it1).op(i));
                                         }
                                         solutions<<");"<<endl;
+
+                                        solNum = solNum + 1;
 
                                         solutionClt[solutionClt.size()-1].append(temList);
                                     }
@@ -424,13 +428,14 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
                             {
                                 if(!is_a<lst>(*it1))
                                 {
-                                    solutions << dpndt_var << " = "  << solu_form_F_subs <<";"<< endl;
+                                    solutions<<"solution #"<<solNum<<"  " << dpndt_var << " = "  << solu_form_F_subs <<";"<< endl;
+                                    solNum = solNum + 1;
 
                                     solutionClt[solutionClt.size()-1].append(dpndt_var  ==  solu_form_F_subs);
                                 }
                                 else
                                 {
-                                    solutions << dpndt_var << " = "  << solu_form_F_subs <<" (with condition(s) ";
+                                    solutions<<"solution #"<<solNum<<"  " << dpndt_var << " = "  << solu_form_F_subs <<" (with condition(s) ";
 
                                     lst temList = {};
                                     temList.append(dpndt_var  ==  solu_form_F_subs);
@@ -441,6 +446,7 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
                                         temList.append((*it1).op(i));
                                     }
                                     solutions<<");"<<endl;
+                                    solNum = solNum + 1;
 
                                     solutionClt[solutionClt.size()-1].append(temList);
 
@@ -477,7 +483,7 @@ int F_expans::operator()(const ex diffeq, const ex dpndt_varChng, const ex dpndt
                         }
                         if(!isUndefine)
                         {
[…]
