### 2021-11-17 2e09fde
add checkSolu() function
