+++ src/desolve.cpp
@@ -27,7 +27,7 @@ ex chi, xi;
 
 ex NValue = 0;
 
-lst twcPhase={}, degAcoeff={}, paraInDiffSolve = {};
+lst twcPhase= {}, degAcoeff={}, paraInDiffSolve = {};
 
 
 
@@ -670,7 +670,7 @@ int desolve(const ex& diffeq, const lst& dpndt_vars, const int& method, bool tes
     {
         cout <<"Evaluation stop: Please provide highest positive integer delta of 1st order NLODE in 'degAcoeff';"<<endl;
         solutions << "Evaluation stop: Please provide highest positive integer delta of 1st order NLODE in 'degAcoeff';"<<endl;
-        writetofile(solutions);
+        writetofile(solutions, dpndt_vars.op(0));
         #ifdef GiNaCDE_gui
         gtk_statusbar_push (GTK_STATUSBAR(status_bar), 0, "Evaluation stop: Please provide highest positive integer delta of 1st order NLODE in 'degAcoeff';");
         //gtk_show_uri(gdk_screen_get_default(),&CurrentPath[0],GDK_CURRENT_TIME,NULL);
@@ -682,7 +682,7 @@ int desolve(const ex& diffeq, const lst& dpndt_vars, const int& method, bool tes
     {
         cout <<"Evaluation stop: delta of 1st order NLODE in 'degAcoeff' should be positive integer;"<<endl;
         solutions << "Evaluation stop: delta of 1st order NLODE in 'degAcoeff' should be positive integer;"<<endl;
-        writetofile(solutions);
+        writetofile(solutions, dpndt_vars.op(0));
         #ifdef GiNaCDE_gui
         gtk_statusbar_push (GTK_STATUSBAR(status_bar), 0, "Evaluation stop: delta of 1st order NLODE in 'degAcoeff' should be positive integer;");
         //gtk_show_uri(gdk_screen_get_default(),&CurrentPath[0],GDK_CURRENT_TIME,NULL);
@@ -694,7 +694,7 @@ int desolve(const ex& diffeq, const lst& dpndt_vars, const int& method, bool tes
     {
         cout <<"Evaluation stop: Please provide all coefficients of 1st order NLODE in 'degAcoeff';"<<endl;
         solutions << "Evaluation stop: Please provide all coefficients of 1st order NLODE in 'degAcoeff';"<<endl;
-        writetofile(solutions);
+        writetofile(solutions, dpndt_vars.op(0));
         #ifdef GiNaCDE_gui
         gtk_statusbar_push (GTK_STATUSBAR(status_bar), 0, "Evaluation stop: Please provide all coefficients of 1st order NLODE in 'degAcoeff';");
         //gtk_show_uri(gdk_screen_get_default(),&CurrentPath[0],GDK_CURRENT_TIME,NULL);
@@ -706,7 +706,7 @@ int desolve(const ex& diffeq, const lst& dpndt_vars, const int& method, bool tes
     {
         cout <<"Evaluation stop: Minimum one part (positivePart or negativePart) should be \"yes\";"<<endl;
         solutions << "Evaluation stop: Minimum one part (positivePart or negativePart) should be \"yes\";"<<endl;
-        writetofile(solutions);
+        writetofile(solutions, dpndt_vars.op(0));
         return -1;
     }
 
@@ -714,7 +714,7 @@ int desolve(const ex& diffeq, const lst& dpndt_vars, const int& method, bool tes
     {
         cout <<"Evaluation stop: For FIM method, value of N is either 1 or 2';"<<endl;
         solutions << "Evaluation stop: For FIM method, value of N is either 1 or 2';"<<endl;
[…]
