### 2021-11-17 398743b
add ginac output format to GUI and gtools
