+++ tools/GiNaCDE_gui.cpp
@@ -105,7 +105,7 @@ static void entry_insert_integer(GtkEditable *entry, char *New, gint len, gpoint
 static void deg_okbutton_clicked(GtkWidget* okbutton,gpointer dialog)
 {
     odedegs = gtk_entry_get_text(GTK_ENTRY(nlodeDeg));
-    
+
     if(odedegs!="")
         gtk_widget_destroy(GTK_WIDGET(dialog));
 }
@@ -121,7 +121,7 @@ static void Acoeff_okbutton_clicked(GtkWidget* okbutton,gpointer dialog)
 
     if(odedegsStr=="")
     {
-        isValid = false; 
+        isValid = false;
 
         /*char *str = g_strdup_printf ("<span font=\"14\" color=\"red\">"
                                "<b>\t\tRed: %d</b>"
@@ -358,7 +358,7 @@ static int on_clicked_evaluatebtn(GtkWidget *_evaluatebtn, gpointer user_data)
                    "%s"
                  "</span>",
                  gtk_label_get_text(GTK_LABEL(dpndtVarLbl)));
-        gtk_label_set_markup(GTK_LABEL(dpndtVarLbl), strLbl);        
+        gtk_label_set_markup(GTK_LABEL(dpndtVarLbl), strLbl);
         //gtk_widget_modify_fg(dpndtVarLbl, GTK_STATE_NORMAL, &gtk_color);
         isValid = false;
     }
@@ -375,7 +375,7 @@ static int on_clicked_evaluatebtn(GtkWidget *_evaluatebtn, gpointer user_data)
                        "%s"
                      "</span>",
                      gtk_label_get_text(GTK_LABEL(dpndtVarLbl)));
-            gtk_label_set_markup(GTK_LABEL(dpndtVarLbl), strLbl);              
+            gtk_label_set_markup(GTK_LABEL(dpndtVarLbl), strLbl);
             //gtk_widget_modify_fg(dpndtVarLbl, GTK_STATE_NORMAL, &gtk_color);
             isValid = false;
         }
@@ -387,7 +387,7 @@ static int on_clicked_evaluatebtn(GtkWidget *_evaluatebtn, gpointer user_data)
                    "%s"
                  "</span>",
                  gtk_label_get_text(GTK_LABEL(indpndtVarLbl)));
-        gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);          
+        gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);
         //gtk_widget_modify_fg(indpndtVarLbl, GTK_STATE_NORMAL, &gtk_color);
         isValid = false;
     }
@@ -415,7 +415,7 @@ static int on_clicked_evaluatebtn(GtkWidget *_evaluatebtn, gpointer user_data)
                            "%s"
                          "</span>",
                          gtk_label_get_text(GTK_LABEL(indpndtVarLbl)));
-                    gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);                 
+                    gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);
                     //gtk_widget_modify_fg(indpndtVarLbl, GTK_STATE_NORMAL, &gtk_color);
                     isValid = false;
                 }
@@ -434,7 +434,7 @@ static int on_clicked_evaluatebtn(GtkWidget *_evaluatebtn, gpointer user_data)
                        "%s"
                      "</span>",
                      gtk_label_get_text(GTK_LABEL(indpndtVarLbl)));
-            gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);             
+            gtk_label_set_markup(GTK_LABEL(indpndtVarLbl), strLbl);
[…]
