### 2021-11-18 82bab8d
update checkSolu() function
