+++ doc/documentation.tex
@@ -1,4 +1,4 @@
-\documentclass[prd,aps,floats,showkeys,nofootinbib,notitlepage]{revtex4}
+\documentclass[prd,aps,floats,showkeys,nofootinbib,notitlepage]{revtex4-2}
 \usepackage{amsmath}
 \usepackage{amssymb}
 \usepackage{amsbsy}
@@ -22,8 +22,8 @@
 \renewcommand{\thesubsection}{\thesection.\arabic{subsection}}
 \renewcommand{\thesubsubsection}{\thesubsection.\arabic{subsubsection}}
 
+
 \begin{document}
-	
 	\begin{titlepage}
 		%\vspace*{3cm}
 		\begin{center}
@@ -52,7 +52,7 @@
 	
 	%In this short tutorial, we have described the following:
 	
-	%In Sec. \ref{sec:compile}, the compiling and installing procedures of the GiNaCDE library have been explained. In Sec. \ref{sec:command}, we present some useful commands which need to solve NLPDEs by the GiNaCDE library. In Sec. \ref{sec:examples}, GiNaCDE library has been applied on some examples to derive their solutions, and we can know the techniques to use GiNaCDE with C++ code. In Sec. \ref{sec:furexam}  we have applied the library on different types of NLPDEs to provide the usage description of the library in detail. In Sec. \ref{sec:furexam} we have also shown some new exact solutions obtained in some NLPDEs. In Sec. \ref{gui}, we give a brief discussion about the GUI version of GiNaCDE.
+	% In Sec. \ref{sec:compile}, the compiling and installing procedures of the GiNaCDE library have been explained. In Sec. \ref{sec:command}, we present some useful commands which need to solve NLPDEs by the GiNaCDE library. In Sec. \ref{sec:examples}, GiNaCDE library has been applied on some examples to derive their solutions, and we can know the techniques to use GiNaCDE with C++ code. In Sec. \ref{sec:furexam}  we have applied the library on different types of NLPDEs to provide the usage description of the library in detail. In Sec. \ref{sec:furexam} we have also shown some new exact solutions obtained in some NLPDEs. In Sec. \ref{gui}, we give a brief discussion about the GUI version of GiNaCDE.
 	\section{Problem}
 	\par Let us consider an NLPDE with independent variables $t,x_1,x_2, \ldots x_m=\mathbf{X}$ and dependent variable $u$ in the following general form
 	\begin{equation}\label{geneq}
@@ -311,7 +311,7 @@
 	
 	
 	\section{Software Implementation}\label{sec:imple}
-	We have implemented the algorithms described in Secs. \ref{sec:Fexpn}, \ref{sec:mF} and \ref{sec:fim}, into GiNaCDE. GiNaCDE is a C++ library that is built on a pure C++ symbolic library GiNaC \cite{ginac}. Besides this library version, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI. When we solve differential equations using GiNaCDE GUI, we do not have to write any C++ code, and compilation of any code is not required. This GUI version guides us in each step to obtain the output results. However, a complete guide of GiNaCDE GUI has been provided in the GiNaCDE \href{https://github.com/mithun218/GiNaCDE}{repository}. In both versions of GiNaCDE (library and GUI), the output results are saved in a text file with calculating steps. Output results can be saved in Maple or Mathematica programming language by assigning the C++ macros {\em Maple} or {\em Mathematica} to the programming variable {\em {output}}. 
[…]
