### 2021-11-19 #3885 @mithun218
Hi @peanutfun Thank you for the detailed review and the valuable suggestions. I am working on them.
I went through your suggestion:

>  Improve the automated tests to verify computation results (or add manual instructions on how to do so): Make tests actually check the computation results mithun218/GiNaCDE#3

and  did just as you suggested, more or less [here.](https://github.com/mithun218/GiNaCDE/pull/16)

1. To check solutions using GiNaCDE we have added _ginac_ output format.
2. To verify computation results we have added checkSolu() function (this function is in primary stage, currently it cannot verify some solutions).
3. We have provided some Maple worksheet, Mathematica notebook and cpp files which illustrate how to test the solutions using Maple, Mathematica and GiNaCDE software, respectively.

Please let us know if we can do anything else.
