### 2021-11-20 mithun218/GiNaCDE#3 @mithun218
@peanutfun Thanks for opening this issue. I have implemented it [here.](https://github.com/mithun218/GiNaCDE/pull/16)
