### 2021-11-20 mithun218/GiNaCDE#5 @mithun218
@peanutfun  Thanks for your information. I will add OS-dependent installation instructions in PR #18
