+++ doc/documentation.tex
@@ -311,7 +311,7 @@
 	
 	
 	\section{Software Implementation}\label{sec:imple}
-	We have implemented the algorithms described in Secs. \ref{sec:Fexpn}, \ref{sec:mF} and \ref{sec:fim}, into GiNaCDE. GiNaCDE is a C++ library that is built on a pure C++ symbolic library GiNaC \cite{ginac}. Besides this library version, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI. When we solve differential equations using GiNaCDE GUI, we do not have to write any C++ code, and compilation of any code is not required. This GUI version guides us in each step to obtain the output results. However, a complete guide of GiNaCDE GUI has been provided in the \href{https://github.com/mithun218/GiNaCDE}{GiNaCDE repository}. In both versions of GiNaCDE (library and GUI), the output results are saved in a text file with calculating steps. Output results can be saved in Maple, Mathematica or GiNaC programming language by assigning the C++ macros {\em maple}, {\em mathematica} or {\em ginac} to the programming variable {\em {output}}. 
+	We have implemented the algorithms described in Secs. \ref{sec:Fexpn}, \ref{sec:mF} and \ref{sec:fim}, into GiNaCDE. GiNaCDE is a C++ library that is built on a pure C++ symbolic library GiNaC \cite{ginac}. Besides this library version, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI. When we solve differential equations using GiNaCDE GUI, we do not have to write any C++ code, and compilation of any code is not required. This GUI version guides us in each step to obtain the output results. However, a complete guide of GiNaCDE GUI has been provided \href{https://sourceforge.net/projects/ginacde}{here}. In both versions of GiNaCDE (library and GUI), the output results are saved in a text file with calculating steps. Output results can be saved in Maple, Mathematica or GiNaC programming language by assigning the C++ macros {\em maple}, {\em mathematica} or {\em ginac} to the programming variable {\em {output}}. 
 	
 	F-expansion and mF-expansion methods can be applied to higher-order NLPDEs. But, FIM method is applicable to an NLPDE when its transformed NLODE \eqref{tweq2} is second-order only. Actually, there are no rules to know in advance the appropriate method among three to solve the given NLPDE. 
 	
@@ -626,7 +626,7 @@
 		\frac{{{k_0} + 2{p_1}{k_1} - 3{G_1}{k_1}p_1^2}}{{ - p_1^2 - {p_0} + {G_1}p_1^3}} = \frac{{{G_1}k_1^3}}{{ - 3{G_1}{p_1}k_1^2 + k_1^2}} = \frac{{2{k_1}{G_3} + {k_1}{G_2}}}{{3A - 3{p_1}{G_2}}}.
 	\end{equation}
 	Same conditions were obtained in \cite{complexTwt1}. Therefore GiNaCDE evaluates the exact analytical solutions of imaginary part only. 
-	We get exact solutions of the NLPDE \eqref{nlseWtKerr} with the F-expansion and mF-expansion methods. FIM is unable to solve the NLPDE \eqref{nlseWtKerr}, because the NLPDE \eqref{nlseWtKerr} in terms of traveling-wave coordinate $\xi$ is not a second-order NLODE. The output results obtained from running the above program have been provided under the examples/ folder of GiNaCDE \href{https://github.com/mithun218/GiNaCDE}{repository} with file names \textit{kerrNLS\_Fexp.txt}, \textit{kerrNLS\_mF.txt} and \textit{kerrNLS\_FIM.txt}.  
[…]
