+++ doc/documentation.tex
@@ -311,7 +311,7 @@
 	
 	
 	\section{Software Implementation}\label{sec:imple}
-	We have implemented the algorithms described in Secs. \ref{sec:Fexpn}, \ref{sec:mF} and \ref{sec:fim}, into GiNaCDE. GiNaCDE is a C++ library that is built on a pure C++ symbolic library GiNaC \cite{ginac}. Besides this library version, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI. When we solve differential equations using GiNaCDE GUI, we do not have to write any C++ code, and compilation of any code is not required. This GUI version guides us in each step to obtain the output results. However, a complete guide of GiNaCDE GUI has been provided \href{https://sourceforge.net/projects/ginacde}{here}. In both versions of GiNaCDE (library and GUI), the output results are saved in a text file with calculating steps. Output results can be saved in Maple, Mathematica or GiNaC programming language by assigning the C++ macros {\em maple}, {\em mathematica} or {\em ginac} to the programming variable {\em {output}}. 
+	We have implemented the algorithms described in Secs. \ref{sec:Fexpn}, \ref{sec:mF} and \ref{sec:fim}, into GiNaCDE. GiNaCDE is a C++ library that is built on a pure C++ symbolic library GiNaC \cite{ginac}. Besides this library version, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI. When we solve differential equations using GiNaCDE GUI, we do not have to write any C++ code, and compilation of any code is not required. This GUI version guides us in each step to obtain the output results. However, a complete guide of GiNaCDE GUI has been provided in the \href{https://github.com/mithun218/GiNaCDE}{GiNaCDE repository}. In both versions of GiNaCDE (library and GUI), the output results are saved in a text file with calculating steps. Output results can be saved in Maple, Mathematica or GiNaC programming language by assigning the C++ macros {\em maple}, {\em mathematica} or {\em ginac} to the programming variable {\em {output}}. 
 	
 	F-expansion and mF-expansion methods can be applied to higher-order NLPDEs. But, FIM method is applicable to an NLPDE when its transformed NLODE \eqref{tweq2} is second-order only. Actually, there are no rules to know in advance the appropriate method among three to solve the given NLPDE. 
 	
@@ -510,7 +510,6 @@
 	Finally, the equation is solved using the F-expansion method by the command {\em desolve} in line 12, and all the output results with calculating steps are saved in the file \emph{NLS{\_}Fexp.txt}. 
 	\par The Command in line 15 solve Eq. \eqref{nls} by {\em FIM} method. 
 	Output results with calculating steps are saved in the file \emph{NLS\_FIM.txt}. 
-	We have supplied the output files \emph{NLS{\_}Fexp.txt} and \emph{NLS\_FIM.txt} in the examples\ folder of GiNaCDE \href{https://github.com/mithun218/GiNaCDE}{repository} separately.
 	\subsection{KdVB Equation}
 	%{\em \textbf{Example 2:}}
 	For solving well-known KdVB equation %\cite{kdvb}
@@ -560,7 +559,6 @@
 	Here positive and negative, both parts are retained in the solutions using the definitions in lines 8 and 9. In line 10, we have not supplied the parameters $p,q$ appearing in input-NLPDE \eqref{KdVB}.
[…]
