+++ doc/documentation.tex
@@ -616,7 +616,7 @@
 			\hline
 		\end{tabular}
 	\end{center}
-    The complete C++ code is given in the examples/ folder of the GiNaCDE \href{https://github.com/mithun218/GiNaCDE}{repository} with a file name \emph{kerrNLS.cpp}.
+    The complete C++ code is given in the examples/ folder of the \href{https://github.com/mithun218/GiNaCDE}{GiNaCDE repository} with a file name \emph{kerrNLS.cpp}.
 	
 	
 	In all methods, the software substitutes the traveling-wave solution \eqref{nls_twtrans} into Eq. \eqref{nlseWtKerr}, and separates the real part and the imaginary part following the step F1. The imaginary part is integrated one time and the software assigns the constant of integration to zero. Then the algebraic expressions of real part and imaginary part are compared and the software detects that they are same equations subject to the following conditions
