### 2021-11-21 #3885 @mithun218
Hi @peanutfun I went through your suggestion:

> Simplify installation instructions for Linux, and possibly add instructions for macOS and Windows: Add OS-dependent installation instructions mithun218/GiNaCDE#5

and did just as you suggested, more or less [here](https://github.com/mithun218/GiNaCDE/pull/18) and [here](https://github.com/mithun218/GiNaCDE/pull/19).

1. I have updated README and documentation files as you suggested in [this](https://github.com/mithun218/GiNaCDE/issues/5) issue.
2. I have removed the tar balls as proposed in [this](https://github.com/mithun218/GiNaCDE/issues/6) issue.

Please let us know if we can do anything more.
