### 2021-11-21 mithun218/GiNaCDE#8 @mithun218
@peanutfun  Thank you for the great suggestions! I will make all of the changes you suggested!
