### 2021-11-21 #3885 @mithun218
Hi @peanutfun I went through your suggestion:

>  Add instructions on how to use GiNaCDE in downstream modules: Support and advertise usage in downstream build systems mithun218/GiNaCDE#8

and did just as you suggested, more or less [here](https://github.com/mithun218/GiNaCDE/pull/20).

1. I have updated README file as you proposed in [this](https://github.com/mithun218/GiNaCDE/issues/8) issue.

Please let us know if we can do anything more.
