+++ JOSS-paper/paper.md
@@ -20,32 +20,26 @@ bibliography: paper.bib
 
 # Summary
 
-GiNaCDE is a free and open-source C++ library, which is an NLPDE (Nonlinear Partial Differential Equation) and NLODE (Nonlinear Ordinary Differential Equation) solver based on the high-performance F-expansion [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn] and First Integral Methods (FIM) [@fim0][@mirza][@complexTwt1]. It unify and gather various researches on F-expansion method and FIM [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion (mF-expansion) methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific community fields by applying any one of the three available methods (F-expansion, modified F-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
+GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on the high-performance F-expansion [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234], modified F-expansion [@modfexpn] and First Integral Methods (FIM) [@fim0][@mirza][@complexTwt1]. It unify and gather various researches on F-expansion method and FIM [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion (mF-expansion) methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific community fields by applying any one of the three available methods (F-expansion, modified F-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
 
 The library has been designed to solve the NLPDEs, which have the following general form
 \begin{equation}\label{geneq}
     F\left( {\alpha_i,u,u_t,u_{x_1},u_{x_1} \ldots ,u_{x_m},u_{tt},u_{t{x_1}},u_{t{x_2}}, \ldots ,u_{t{x_m}},u_{{x_1}{x_1}},u_{{x_1}{x_2}} \ldots ,u_{{x_1}{x_m}} \ldots } \right) = 0,
 \end{equation}
-where $t,x_1,x_2, \ldots x_m$ are independent variables, $u=u(t,x_1,x_2, \ldots x_m)$ is dependent variable, $\alpha_i(i=1,2,\ldots,n)$ are the parameters. Here $F$ must be a polynomial about $u$ and its derivatives. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific community fields by applying any one of the three available methods (F-expansion, modified F-expansion, and first integral method). The primary intention of GiNaCDE is to facilitate the development and validation of new exact analytical solutions of NLPDEs and NLODEs. 
[…]
