+++ JOSS-paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on the high-performance F-expansion [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234], modified F-expansion [@modfexpn] and First Integral Methods (FIM) [@fim0][@mirza][@complexTwt1]. It unify and gather various researches on F-expansion method and FIM [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion (mF-expansion) methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific community fields by applying any one of the three available methods (F-expansion, modified F-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
+GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on the high-performance F-expansion [@fexpn024], modified F-expansion [@modfexpn] and First Integral Methods (FIM) [@fim0]. It unify and gather various researches on F-expansion method and FIM [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion (mF-expansion) methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific community fields by applying any one of the three available methods (F-expansion, modified F-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
 
 The library has been designed to solve the NLPDEs, which have the following general form
 \begin{equation}\label{geneq}
@@ -30,7 +30,7 @@ where $t,x_1,x_2, \ldots x_m$ are independent variables, $u=u(t,x_1,x_2, \ldots 
 
 # Statement of need 
 
-The NLODEs and NLPDEs play an important role in the theoretical sciences to explain many nonlinear phenomena in various fields of science, such as biology, chemistry, engineer, solid-state physics, plasma physics, optical fibers. The exact (closed-form) traveling-wave solutions of such NLPDEs give much extra information, which helps us to study the result more deeply. The knowledge of closed-form solutions of NLODEs and NLPDEs helps to test the degree of accuracy of numerical solvers and also facilitates stability analysis. In the past few decades, many powerful methods have been presented for seeking exact solutions of NLPDEs, such as F-expansion method [@fexpn024] and first integral method [@fim0]. 
+The NLODEs and NLPDEs play an important role in the theoretical sciences to explain many nonlinear phenomena in various fields of science, such as biology, chemistry, engineer, solid-state physics, plasma physics, optical fibers. The exact (closed-form) traveling-wave solutions of such NLPDEs give much extra information, which helps us to study the result more deeply. The knowledge of closed-form solutions of NLODEs and NLPDEs helps to test the degree of accuracy of numerical solvers and also facilitates stability analysis. In the past few decades, many powerful methods have been presented for seeking exact solutions of NLPDEs, such as F-expansion method and first integral method. 
[…]
