### 2021-11-22 mithun218/GiNaCDE#12 @mithun218
Thanks @peanutfun for the thoughtful review! and the suggestions!

Below, we provide a point-by-point response explaining how we have addressed each of your suggestions:

1.

> I have noticed some grammatical and spelling errors in the paper. Since English is not my mother tongue, I suggest you use a spell and grammar checker on the paper. You could use https://www.grammarly.com/, MS Word, or some editor plugin of choice.

We have checked the paper thoroughly, and have corrected all the grammatical errors and misspellings.

2.

> The "Algorithms" section seems redundant to me. You mention all methods and references in previous sections already. Your explanation of the formation of the methods from this section would fit well into the first paragraph of the "Statement of Need" section.

Thank you for raising this point. We have moved the "Algorithms" section into the first paragraph of the "Statement of Need" section.

3.

> The references for F-expansion ([@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn]) and first integral methods ([@fim0][@mirza][@complexTwt1]) appear several times throughout the paper and even several times in single sections. Please avoid repeated citations of the same paper and cite a source only once per section. Also, please clarify which references explain the modified F-expansion, and which the "regular" F-expansion method.

We have removed the repetition of the same references as far as possible. In some cases, we have to repeatedly use the same references to clarify the source of the statement. We have clarified which references explain the modified F-expansion, and which the "regular" F-expansion method.

4.

> Avoid "etc." after listings and use "such as" or "e.g." before listings instead.

We agree with you. We have done as you suggested.

5.
>Avoid "which is" here, as GiNaCDE being a solver is more important than being a C++ library. Suggestion:

You are correct. We have done as you suggested.

6.

> Please add a suitable reference for this statement.

Actually, we have stated this statement from our observations. So instead of adding a reference to this statement, we have changed the statement as follows:
“On the other hand, we have observed that F-expansion, mF-expansion, and FIM are different kinds of methods,……..”

7.

> This paragraph would fit better into the summary. Instead of giving a number of equations, I suggest listing (only) the equations given in the examples in the repository. Also, please state the unabbreviated names of these equations instead of acronyms.

We have moved this paragraph into the summary, and we have used unabbreviated names of the NLPDEs.

8.

> It is unclear here if this time refers to solving one or all equations you mention, and if compile times are included. Further, executable run times are highly dependent on the system hardware. Please remove this claim altogether or make it more precise, e.g. by benchmarking with similar methods from other packages.

Thank you for raising this point. We have removed this claim.
