### 2021-11-22 #3885 @mithun218
Hi @peanutfun I went through your suggestion:
> Incorporate or conclusively reject suggestions to the JOSS paper: Improve the JOSS paper mithun218/GiNaCDE#12

and did just as you suggested, more or less [here](https://github.com/mithun218/GiNaCDE/pull/21).

Please let us know if we can do anything more.
