### 2021-11-24 mithun218/GiNaCDE#3 @peanutfun
@mithun218 thanks for implementing the `checkSolu` function! However, you did not add it to the existing tests `test1.cpp` and `test2.cpp`, and vice versa, you did not add the new tests to the CMake test configuration. Please add the new tests so that they are executed when running `ctest`, and add the `checkSolu` queries to the original test source files as well.
