### 2021-11-24 mithun218/GiNaCDE#8 @peanutfun
Thank you! You now mention the issue with the `#include` path of GiNaCDE in the `README.md`, but I think this is a bit confusing. I suggest you change the include path in `ginacde.pc.in` to `@CMAKE_INSTALL_PREFIX@/include/`, so that `#include <GiNaCDE/GiNaCDE.h>` is always correct. Then the source files do not need to change depending on the way GiNaCDE is installed or included into a downstream project.
