### 2021-11-24 mithun218/GiNaCDE#12 @peanutfun
Thanks a lot for going through my suggestions, and for responding so thoroughly! ☺️
