### 2021-11-24 mithun218/GiNaCDE#6 @peanutfun
Thank you! Sourceforge is a great place to store the GUI 👍
