### 2021-11-24 mithun218/GiNaCDE#3 @mithun218
@peanutfun  thanks for your quick response.
I am sorry to say that currently `checkSolu` function is unable to check some solutions returned by GiNaCDE due to some simplification problems in GiNaCDE (we hope we can get out of this problem in the future release). So if we use `checkSolu` function to the existing tests `test1.cpp` and `test2.cpp`, it may report wrong test results for a correct solution. As a result, we may miss some important solutions of input NLPDE or NLODE.  For this reason, we have not used the `checkSolu` function in the test files, and GiNaCDE also reports the solutions without using the `checkSolu` function.

However, in the README file, we have explained the procedure for verifying the solutions. Here we have been told to check the solutions manually using `checkSolu` function of GiNaCDE or Maple and Mathematica software  (Here manually means we have to copy the exact solutions with the conditions of solutions (which are also provided in the output .txt file) from the output .txt file ourselves.).
To illustrate the procedures of checking the solutions, we have provided some output text files and the corresponding checking files. These checking files test some solutions reported in the provided text files, and explain how to test the solutions using Maple (Maple 2019), Mathematica (Mathematica 9), and GiNaCDE software.

Finally, to check all the solutions successfully, we recommend to use Maple or Mathematica software instead of using the `checkSolu` function.
