### 2021-11-24 mithun218/GiNaCDE#8 @mithun218
@peanutfun thanks for your great suggestion. I will implement it in a new PR.
