+++ examples/NLS.cpp
@@ -7,7 +7,7 @@
 
 
 
-#include <GiNaCDE.h>
+#include <GiNaCDE/GiNaCDE.h>
 
 
 int main()
