+++ GiNaCDE.pc.in
@@ -1,7 +1,7 @@
 prefix=@CMAKE_INSTALL_PREFIX@
 exec_prefix=@CMAKE_INSTALL_PREFIX@/bin
 libdir=@CMAKE_INSTALL_PREFIX@/lib
-includedir=@CMAKE_INSTALL_PREFIX@/include/GiNaCDE
+includedir=@CMAKE_INSTALL_PREFIX@/include/
 
 Name: @PROJECT_FULLNAME@
 Description: @PROJECT_DESCRIPTION@
