### 2021-11-24 mithun218/GiNaCDE#8 @mithun218
I have implemented it [here](https://github.com/mithun218/GiNaCDE/pull/22).
