### 2021-11-26 mithun218/GiNaCDE#3 @peanutfun
> I am sorry to say that currently checkSolu function is unable to check some solutions returned by GiNaCDE due to some simplification problems in GiNaCDE [...]

This is alright. However, a test that does not perform any checks is not a test. As I initially mentioned, it is not required to have a universal function that is able to check all GiNaCDE solutions. For the test cases where `checkSolu` fails, you can still add the expected solution as GiNaC expression manually and check if this expression and the solution found by GiNaCDE yield the same results for a set of pre-defined parameters.

Also, this is independent from adding the new test cases that include the `checkSolu` function to the executables and tests registered in CMake.
