+++ test/test1.cpp
@@ -3,8 +3,7 @@
  *
  *   Program to test GiNaCDE library. This solve following odes
         i.   linear ode - the damped harmonic oscillator.
-        ii.  nonlinear ode - modified Painlev-Ince equation
-        iii. Kudryashov-Sinelshchikov equation */
+        ii.  nonlinear ode - modified Painlev-Ince equation */
 
 
 
@@ -15,7 +14,10 @@
  {
     const ex u=reader("u"),x=reader("x"),a=reader("a"),b=reader("b"),
              c=reader("c"),w=reader("w"),A_0=reader("A_0"),A_1=reader("A_1"),A_2=reader("A_2");
-    ex ode;
+    ex ode,res;
+    size_t solu_num1,solu_num2;
+    stringstream diffStr,algSoluStr,diffSoluStr;
+    string str;
 
     depend(u, {x});
 
@@ -25,6 +27,26 @@
     output = maple;// Outputs are saved in maple format;
     filename = "damped_FIM.txt";
     desolve(ode,{u},FIM,true);
+    /* Checking all solutions*/
+    diffStr.str("");
+    diffStr<<ode;
+    solu_num1 = sizeof(solutionClt);
+    for(size_t i=1;i<solu_num1;i++)
+    {
+        algSoluStr.str("");
+        algSoluStr<<solutionClt[i][0];
+        solu_num2 = sizeof(solutionClt[i]);
+        for(size_t j=1;j<solu_num2;j++)
+        {
+            diffSoluStr.str("");
+            diffSoluStr<<solutionClt[i][j];
+            res = checkSolu(diffStr.str(),diffSoluStr.str(),algSoluStr.str());
+            if(res!=_ex0)
+                return -1;
+        }
+
+    }
+
 
 
     ode = Diff(u,x,2) + a*u*Diff(u,x,1) + b*u*u*u; // modified Painlev-Ince equation
@@ -34,17 +56,25 @@
     filename = "Painlev_FIMextravar.txt";
     paraInDiffSolve = lst{a,b};
     desolve(ode,{u},FIM,true);
+    /* Checking all solutions*/
+    diffStr.str("");
+    diffStr<<ode;
+    solu_num1 = sizeof(solutionClt);
+    for(size_t i=1;i<solu_num1;i++)
+    {
+        algSoluStr.str("");
+        algSoluStr<<solutionClt[i][0];
+        solu_num2 = sizeof(solutionClt[i]);
+        for(size_t j=1;j<solu_num2;j++)
+        {
+            diffSoluStr.str("");
+            diffSoluStr<<solutionClt[i][j];
+            res = checkSolu(diffStr.str(),diffSoluStr.str(),algSoluStr.str());
+            if(res!=_ex0)
+                return -1;
+        }
 
-
-    /*ode = c*Diff(u,x,1)-a*(1-u)*Diff(u,x,1)-Diff(u,x,3)+Diff((1-u)*Diff(u,x,2),x,1)-b*Diff(u,x,1)*Diff(u,x,2);  // Kudryashov-Sinelshchikov equation
-    output = maple;
-    NValue = 1;
-    degAcoeff = lst{2,1,1,A_2};
-    filename = "kudryashov_mF.txt";
-    paraInDiffSolve = lst{a,b,c};
-    desolve(ode, {u}, mF_expansion,true);*/
-
-
+    }
 
     return 0;
 
