### 2021-11-26 7940ebb
add some tests to CMakeLists
