+++ JOSS-paper/paper.md
@@ -32,7 +32,7 @@ nonlinearity terms, higher-derivative terms, and complex NLPDE. Some of them are
 # Statement of need 
 
 The NLODEs and NLPDEs play an important role in the theoretical sciences to explain many nonlinear phenomena in various fields of science, such as biology, chemistry, engineer, solid-state physics, plasma physics, optical fibers. The exact (closed-form) traveling-wave solutions of such NLPDEs give much extra information, which helps us to study the result more deeply. The knowledge of closed-form solutions of NLODEs and NLPDEs helps to test the degree of accuracy of numerical solvers and also facilitates stability analysis. In the past few decades, many powerful methods have been presented to seek exact solutions of NLPDEs, such as F-expansion method and first integral method. 
-The F-expansion method was first proposed by Zhou et al. [@fexpn024]. The first integral method was first introduced by Feng [@fim0] in solving the Compound Burgers-KdV Equation, which is based on the ring theory of commutative algebra. Later, these methods have been further improved in a number of research works. Following some research works [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1], GiNaCDE intends to gather and unify the possible combinations of the different revised versions of these methods for improvements of solution procedures. In this context, on the base of the aforementioned research works of these methods, we have presented the high-performance algorithms of F-expansion, modified F-expansion, and first integral methods in the documentation[^1] file. The GiNaCDE software uses these algorithms to solve the NLPDEs and NLODEs.
+The F-expansion method was first proposed by Zhou et al. [@fexpn024]. The first integral method was first introduced by Feng [@fim0] in solving the Compound Burgers-KdV Equation, which is based on the ring theory of commutative algebra. Later, these methods have been further improved in a number of research works. Following some research works [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1], GiNaCDE intends to gather and unify the possible combinations of the different revised versions of these methods for improvements of solution procedures. In this context, on the basis of the aforementioned research works of these methods, we have presented the high-performance algorithms of F-expansion, modified F-expansion, and first integral methods in the documentation[^1] file. The GiNaCDE software uses these algorithms to solve the NLPDEs and NLODEs.
 
 In order to solve the NLPDEs, many computer packages are available.
 In 1996, Parkes and Duffy [@atfm] had implemented tanh-expansion in their Mathematica package ATFM. Later complete implementation of tanh-expansion has been done by Li and Liu (2002) [@rath] designing the Maple package RATH. Baldwin et al. [@pdespclpkg1] have developed the Mathematica package *PDESpecialSolutions.m* which admits polynomial solutions in tanh, sech, combinations thereof, JacobiSN, JacobiCN. RAEEM [@raeem] is one of the most popular packages written in the Maple programming language, which is a comprehensive and complete implementation of some powerful methods such as the tanh-method, the extended tanh-method, the Jacobi elliptic function method, and the elliptic equation method. One can note that most packages have been developed within commercially available software frameworks, such as Maple and Mathematica. Besides this, all these computer packages have implemented the function-expansion methods. One serious drawback of the function-expansion method is that the solutions which contain functions other than some specific type of functions, such as tanh, sech, JacobiSN, JacobiCN, are not obtained. Additionally, the non-polynomial forms of these particular functions are not obtained also. On the other hand, we have observed that F-expansion, mF-expansion, and FIM are different kinds of methods, which can overcome the limitations of the function-expansion method.   
