### 2021-11-27 #3885 @mithun218
> joss paper correction mithun218/GiNaCDE#24

I again went through the paper, and I changed only one word.
The change is:
on the base of ⇾ on the basis of
