### 2021-11-29 c3b8e46
check degAcoeff initialization
