### 2021-11-30 9d49fc1
move and update checkSolu test files
