### 2021-11-30 0916e70
add checkSolu files to make examples
