### 2021-12-06 mithun218/GiNaCDE#4 @mithun218
@peanutfun Thanks a lot for your valuable proposal. I also think GiNaCDE would attract a larger audience by supporting an additional output format that does not require a software subscription. Recently, we have added the output format `ginac` in [this ](https://github.com/mithun218/GiNaCDE/pull/16)PR. As a result, now one can check the solutions in GiNaCDE also. We will soon add some non-proprietary output formats to GiNaCDE, such as Python, LaTex.
