### 2021-12-06 mithun218/GiNaCDE#11 @mithun218
@peanutfun Thanks a lot for your great suggestion! I will work soon on your proposal.
