### 2021-12-06 mithun218/GiNaCDE#10 @mithun218
@peanutfun thanks for your suggestion. I am sorry to say that currently, I am entirely ignorant of Doxygen. From your explanation, it looks excellent to use Doxygen in GiNaCDE. I hope that I will be able to use Doxygen to GiNaCDE very soon.
