### 2021-12-06 mithun218/GiNaCDE#2 @mithun218
@peanutfun Thanks for your valuable observations. It is correct that GiNaCDE does not return the results in GiNaC expression ex. But, GiNaCDE returns its results through the variables `solutionClt` and `constraints`. One can use the results collected by the variables `solutionClt` and `constraints`  for further computations within a program.
Please let us know, is this enough?
