### 2021-12-08 #3885 @mithun218
@peanutfun  I have addressed all the issues you referenced  [here](https://github.com/openjournals/joss-reviews/issues/3885#issuecomment-965388988) . Please review and if I have missed anything please let me know.
