### 2021-12-08 mithun218/GiNaCDE#3 @peanutfun
@mithun218, thank you for implementing my proposals. I did not realize you already finished working on it. I just checked out the latest `master` and had a look at the tests. In general, they look good now, but it seems to me like the `checkSolu` function is not working correctly. In the `test/checkSolu_Painlev_FIMextravar.cpp` file, I changed the solution string, but the `checkSolu` function still returns the same value, and the test passes although the solution is wrong.  Can you give this another look? Here is an example:

```c++
#include <GiNaCDE.h>

int main()
{
    const string DE = "Diff(u,x,2)+u^3*b+u*Diff(u,x,1)*a";

    // Original solution (works)
    string algebraic_solutions = "{g_0==0,a_01==0,a_00==0,b==-1/2*(a+g_1)*g_1,a_02==1/2*a+1/2*g_1}";
    string solutions = "u = 2*(2*C_+(a+g_1)*x)^(-1)";

    // Wrong solutions that still work
    string solutions_2 = "u =  2*(2*C_+(a+g_1)*x)^(-2)";  // Wrong exponent
    string solutions_3 = "u =  1 + 2*(2*C_+(a+g_1)*x)^(-2)";  // Additional summand

    std::vector<string> solution_vec({solutions, solutions_2, solutions_3});
    for (auto sol : solution_vec) {
        ex ret = checkSolu(DE, sol, algebraic_solutions);
        if(ret!=_ex0)
        {
            // We never end up here!
            return -1;
        }
    }

    return 0;
}
```
