### 2021-12-08 mithun218/GiNaCDE#4 @peanutfun
Thank you for implementing the `ginac` format, I think this was a very important step. If you plan on adding more output formats soon, I propose to leave this issue open such that other users see that this is still under consideration.
