### 2021-12-08 mithun218/GiNaCDE#2 @peanutfun
I'm sorry, I missed these global return variables on my first go through your program. I think they suffice, but it might be worth considering not to have them as global variables, but only local variables that are returned by the `desolve` function. This would work well in conjunction with my proposal in #11. With a class encapsulation, you can define all these variables locally. I'll leave the decision up to you, feel free to close this issue!
