### 2021-12-08 e7ba721
bug fixes of checkSolu function
