+++ test/test2.cpp
@@ -27,37 +27,6 @@ int main()
     depend(u, {t, x, y, z});
 
 
-    pde = Diff(u,t,1)+Diff(u,x,1)+u*Diff(u,x,1)-Diff(Diff(u,x,2),t,1);  // Benjamin-Bona-Mahony equation
-    output = ginac;
-    twcPhase=lst{lst{k_0,k_1},lst{}};
-    degAcoeff = lst{4,0,0,A_2,0,A_4};
-    ASolve = true;
-    positivePart = true;
-    negativePart = false;
-    filename = "benjamin_Fex.txt";
-    desolve(pde, {u}, F_expansion,true);
-    /* Checking all solutions*/
-     diffStr.str("");
-     diffStr<<pde;
-     solu_num1 = (solutionClt).size();
-     for(size_t i=0;i<solu_num1;i++)
-     {
-         algSoluStr.str("");
-         algSoluStr<<solutionClt[i][0];
-         solu_num2 = nops(solutionClt[i]);
-         for(size_t j=1;j<solu_num2;j++)
-         {
-             diffSoluStr.str("");
-             diffSoluStr<<solutionClt[i][j];
-             res = checkSolu(diffStr.str(),diffSoluStr.str(),algSoluStr.str());
-             if(res!=_ex0)
-                 return -1;
-         }
-
-     }
-
-
-
     pde = Diff(u,t,1) - Diff(u,x,2) + u*u*u - u;  // Cahn-Allen equation
     output = ginac;
     twcPhase=lst{lst{k_0,k_1},lst{}};
@@ -87,7 +56,6 @@ int main()
      }
 
 
-
     pde = Diff(u,t,2)+b*u+c*pow(u,3);  // Duffing equation
     output = ginac;
     twcPhase=lst{lst{},lst{}};
@@ -113,36 +81,6 @@ int main()
          }
 
      }
-
-
-
-    pde = Diff(u,t,1)+u*Diff(u,x,1)+Diff(u,x,3)-Diff(u*Diff(u,x,2),x,1)-Diff(u,x,1)*Diff(u,x,2)-Diff(u,x,2)-Diff(u*Diff(u,x,1),x,1);  // Kudryashov-Sinelshchikov  equation
-    output = ginac;
-    twcPhase=lst{lst{k_0,k_1},lst{}};
-    degAcoeff = lst{2,0,A_1,A_2};
-    NValue = 2;
-    filename = "Kudryashov-Sinelshchikov_mF.txt";
-    desolve(pde, {u}, mF_expansion,true);
-    /* Checking all solutions*/
-     diffStr.str("");
-     diffStr<<pde;
-     solu_num1 = (solutionClt).size();
-     for(size_t i=0;i<solu_num1;i++)
-     {
-         algSoluStr.str("");
-         algSoluStr<<solutionClt[i][0];
-         solu_num2 = nops(solutionClt[i]);
-         for(size_t j=1;j<solu_num2;j++)
-         {
-             diffSoluStr.str("");
-             diffSoluStr<<solutionClt[i][j];
-             res = checkSolu(diffStr.str(),diffSoluStr.str(),algSoluStr.str());
-             if(res!=_ex0)
-                 return -1;
-         }
-
-     }
-
 
 
     pde = Diff(u,t,2)-Diff(u,x,2)+Diff(u,t,1)+a*u+b*pow(u,3);  //  Nonlinear Telegraph Equation
