### 2021-12-08 mithun218/GiNaCDE#3 @mithun218
@peanutfun Many many thanks for detecting this problem of the `checkSolu` function. I have noticed the problem and resolved the issue [here](https://github.com/mithun218/GiNaCDE/pull/28). `checkSolu` function now works fine. Please check this.
