### 2021-12-15 #3885 @peanutfun
@mithun218 Thank you for resolving the major issue of automated tests, and the other issues I brought up as critical for the review.

@meg-simula I finished my review and I endorse GiNaCDE for publication.
