### 2021-12-15 mithun218/GiNaCDE#3 @peanutfun
Very good, `checkSolu` now appears to work correctly. Thank you! Together with the automated tests, this resolves the original issue.
