### 2022-01-31 mithun218/GiNaCDE#29 @carstenbauer
Instructions for non-systemwide installation
(This issue is part of my JOSS review, see https://github.com/openjournals/joss-reviews/issues/3885.)

Currently, the instructions focus on systemwide installations of both the dependencies (CLN and GiNaC) as well as GiNaCDE itself. I think it would be very helpful to also mention the necessary extra flags to install the software locally (e.g. into `~/.local` or similar). This is particularly useful for usage on clusters and corporate machines where one typically doesn't have `sudo` rights.

* For CLN and GiNaC the necessary bit is augmenting the `./configure` bit by `--prefix=~/.local`. This *could* be noted here but is also explained in their respective documentation. (optional)
* For GiNaCDE, one needs to add `-DCMAKE_INSTALL_PREFIX:PATH=~/.local` to the `cmake` call, i.e. `cmake -DGINACDE_GUI_BUILD=on -DCMAKE_INSTALL_PREFIX:PATH=~/.local <path-to-source> # generate Makefiles`
