### 2022-01-31 mithun218/GiNaCDE#30 @carstenbauer
Improve documentation for how to run the tests
(This issue is part of my JOSS review, see openjournals/joss-reviews#3885.)

I would like to suggest improvements for the [tests section](https://github.com/mithun218/GiNaCDE#tests) of the README.md.

* mention `make test` besides `ctest`: some people might be more familiar with `make test` so it would be nice to also mention this option for running the tests.
* To avoid potential confusion, make clear that `ctest` or `make test` needs to be run in `build-dir`: the current formulation “The `test` folder contains tests” might (wrongly) suggest that one needs to run `ctest` in this test folder.
