### 2022-01-31 mithun218/GiNaCDE#31 @carstenbauer
Revise structure of README.md
(This issue is part of my JOSS review, see openjournals/joss-reviews#3885.)

**Currently**, the order is as follows:
* Summary
* Features
* External dependency (-> should be "External dependencies" by the way)
* Compiling and installing
* Tests
* Checking the solutions of Diff. Equ.
* Execution
* Output
* Examples
  * Additional notes (-> should probably be moved to the Output section?)
* Documentation
* Example usage
* GUI build (-> I will rename it to "Precompiled GiNaCDE-GUI for Windows" below)
* Contributions and bug reports

I think for better clarity it would be worth reordering things a bit and introducing some hierarchy (different header levels). For example, I suggest the following structure:
* Summary
* Features
* Example usage (-> IMO, it is always good to show an example as early as possible. This helps people to decide whether they should continue reading and install the software in the first place.)
* Installation
  * External dependencies
  * Compiling and installing GiNaCDE
    * Precompiled GiNaCDE-GUI for Windows
  * Automated Tests (-> it's debatable whether this should be a subsection of Installation, but I think I like it here)
* Execution
  * Output
    * Additional notes
* Examples
* Documentation
* Contributions and bug reports

Of course, there is a bit of subjectivity in all of this, which is why I won't mind if you decide to not implement this precise structure.
