### 2022-01-31 #3885 @carstenbauer
Alright, I've finished my review and have completed the checklist in the OP. I think that both the GiNaCDE software as well as the manuscript meets all of the JOSS criteria (in particular after the improvements triggered by @peanutfun's comments). The software has a clear purpose, is well and extensively documented (in particular from a methods point of view) and works as expected (I tested the included examples as well as a simple self-created example). Given that the raised issues below will be addressed soon, I happily endorse GiNaCDE for publication.

(Note that I did not test / review the optional GUI component.)

### Issues related to my review
* [ ] Instructions for non-systemwide installation - https://github.com/mithun218/GiNaCDE/issues/29
* [ ] Improve documentation for how to run the tests - https://github.com/mithun218/GiNaCDE/issues/30
* [ ] Revise structure of README.md - https://github.com/mithun218/GiNaCDE/issues/31
