+++ examples/gardner_Fex.txt
@@ -4,7 +4,7 @@ Equations are written in MATHEMATICA language.
 
 Input equation is: D[u[x,t],{x,3}]+2*a*u[x,t]*D[u[x,t],{x,1}]-3*b*u[x,t]^2*D[u[x,t],{x,1}]+D[u[x,t],{t,1}] = 0;
 The Diff. Equ. is integrable;
-The assigned Value[s] to the integrating Constant[s]-> 
+The assigned Value[s] to the integration Constant[s]-> 
 ic1: 0
 The Diff. Equ. becomes: a*U[xi]^2*k1-b*U[xi]^3*k1+U[xi]*k0+k1^3*D[U[xi],{xi,2}] = 0;
 u[x,t] = U[xi], 
