+++ doc/GiNaCDE_guiTutorial.tex
@@ -1,0 +1,171 @@
+\documentclass[11pt,a4paper,titlepage]{article}
+\usepackage{amsmath}
+\usepackage{amssymb}
+\usepackage{wasysym}
+\usepackage{float}
+\usepackage{latexsym,amsmath,amssymb}
+\usepackage{graphicx}
+\usepackage{epstopdf}
+%\usepackage{subfig}
+\usepackage{slashed}
+\usepackage{bm}
+\usepackage{epsfig}
+\usepackage{cases}
+\usepackage{dcolumn}
+\usepackage{color}
+\usepackage[colorlinks=true,linkcolor=red]{hyperref}
+\usepackage{listings} 
+\usepackage{xcolor} % for setting colors
+\begin{document}
+\begin{titlepage}
+%\vspace*{3cm}
+\begin{center}
+\Huge {A tutorial on GiNaCDE-GUI to solve Differential Equations} 
+\end{center}
+\vspace{3cm}
+\begin{center}
+\Large{GiNaCDE-GUI (V1.0.0)}
+\end{center}
+\vspace{2cm} 
+\begin{center} 
+Mithun Bairagi \\[3pt]  
+\textit{Department of Physics, The University of Burdwan, Golapbag 713104, West Bengal, India} \\ [1cm]
+%\textit{Address} \\ [1cm] 
+\today
+\end{center}
+\end{titlepage}
+ 
+\clearpage
+The usage of GiNaCDE-GUI for solving differential equations is described in this tutorial. The following two examples demonstrate the usage of GiNaCDE-GUI to solve differential equations.
+\section{Example 1}
+Let us consider the Eckhaus equation 
+\begin{equation}\label{eckhaus}
+	i{u_t} + {u_{xx}} + 2{\left( {{{\left| u \right|}^2}} \right)_x}u + {\left| u \right|^4}u = 0.
+\end{equation}
+For solving Eq. \eqref{eckhaus} using GiNaCDE library, the C++ codes are\\
+\begin{verbatim}
+	// eckhaus_FIM.cpp
+	#include <GiNaCDE/GiNaCDE.h>
+	int main()
+	{
+		   const ex u=reader("u"), t=reader("t"), x=reader("x"), 
+		            a=reader("a"), b=reader("b"),k=reader("k");   
+		   depend(u, {t, x});
+		   const ex pde = I*Diff(u,t,1) + Diff(u,x,2) + 2*u*Diff(u*conjugate(u),x,1)
+		                  + u*u*conjugate(u)*conjugate(u)*u;
+		   output = mathematica;  
+		   twcPhase=lst{lst{-2*k*a,k},lst{b,a}};
+		   paraInDiffSolve=lst{};
+		   filename = "eckhaus_FIM.txt";
+		   desolve(pde,{u},FIM);
+		   return 0;
+	}
+\end{verbatim}
+In the following, we display the screenshots of each step when we implement the above C++ code in GiNaCDE-GUI.
+
+\begin{figure}[H]
+	\centering
+	\includegraphics[width=\linewidth]{Ex1step1}
+	\caption{ Step 1. }
+	\label{fig:Ex1step1}
+\end{figure}
+
+\begin{figure}[H]
+	\centering
+	\includegraphics[width=\linewidth]{Ex1step2}
+	\caption{ Step 2. }
+	\label{fig:Ex1step2}
+\end{figure}
+
+\begin{figure}[H]
+	\centering
+	\includegraphics[width=\linewidth]{Ex1step3}
+	\caption{ Step 3. }
+	\label{fig:Ex1step3}
+\end{figure}
+
+\begin{figure}[H]
+	\centering
[…]
