### 2022-01-31 mithun218/GiNaCDE#7 @mithun218
Thank you for your suggestions. I have implemented it [here.](https://github.com/mithun218/GiNaCDE/pull/33)
