### 2022-01-31 9a6afe4
Instructions for non-systemwide installation #29
