### 2022-01-31 09ab155
Improve documentation for how to run the tests #30
