+++ README.md
@@ -12,6 +12,7 @@ F-expansion and modified F-expansion methods can be applied to higher-order NLPD
 However, there is no guarantee that the library always gives the complete solutions of all NLPDEs of the above form. Sometimes, the library may fail to give solutions due to the complexity of the problems.
 
 **N.B.: Currently, GiNaCDE determines the solutions of differential equations assuming all the constant parameters are strictly real and positive.**
+
 ## Features
 Some interesting features of GiNaCDE are
 
@@ -23,113 +24,6 @@ Some interesting features of GiNaCDE are
   * The exact analytical solutions of NLPDEs or NLODES with calculating steps are saved in a text file written in `MAPLE`, `MATHEMATICA`, or `GiNaC` language.
   * It has a friendly Graphical User Interface (GUI).
 
- 
-## External dependency
-GiNaCDE V1.0.0 requires the packages [CLN >= 1.3.4](http://www.ginac.de/CLN/), [GiNaC >= 1.7.6](https://www.ginac.de/archives/) and [GTK+ 3.xx](https://download-fallback.gnome.org/sources/gtk+/3.24/) (this library is optional and is used to build the GUI version of the GiNaCDE library). 
-
-#### For Linux/MacOS machines:
-All the dependencies are available via the most common package managers `APT` on Ubuntu or Debian. Additionally, all dependencies can also be retrieved on `macOS` via the most common package managers [Homebrew](https://brew.sh/) and [MacPorts](https://ports.macports.org/). For example, CLN, GiNaC and GTK3 are installed via APT through
-
-```apt install libcln-dev libginac-dev libgtk-3-dev```
-
-and on Homebrew via
-
-```brew install cln ginac gtk+3```
-
-#### For Windows machines:
-
- We have to install the libraries  [CLN >= 1.3.4](http://www.ginac.de/CLN/), [GiNaC >= 1.7.6](https://www.ginac.de/archives/) using the following commands:  
-```        
-        $ ./configure
-        $ make
-        $ make install
-```
- The library  [GTK+ 3.xx](https://download-fallback.gnome.org/sources/gtk+/3.24/) can be easily installed using [MSYS2](https://msys2.github.io/), which provides a UNIX-like environment for Windows. It provides packages for many software applications and libraries, including the GTK stack. To install GTK3 and its dependencies, open a MSYS2 shell, and run:
-```
-        $ pacman -S mingw-w64-x86_64-gtk3
-```   
-
-## Compiling and installing
-Compilations are done using the tools [CMake](https://cmake.org/download/) `>= 3.1`, pkg-config `>=0.29.2`, and of course, a compiler that supports `>= C++11`.
-We suggest to use the C++ compiler from the GNU compiler collection, `GCC >= 4.9`. GiNaCDE can then be compiled using the commands: 
-```
-     $ mkdir build-dir # generate a separate directory
-     $ cd build-dir
-     $ cmake -DGINACDE_GUI_BUILD=on <path-to-source> # generate Makefiles
-     $ make
-     $ make install
-```
-To install the software locally (e.g. into `~/.local` or similar), one needs to add `-DCMAKE_INSTALL_PREFIX:PATH=~/.local` to the `cmake` call, i.e.
-```
-cmake -DGINACDE_GUI_BUILD=on -DCMAKE_INSTALL_PREFIX:PATH=~/.local <path-to-source> # generate Makefiles
-```
-
-A successful compilation will lead to the creation of libraries, executables of gtools, and GiNaCDE-GUI. If you do not want to build GiNaCDE-GUI, use the following option:
-```
-    -DGINACDE_GUI_BUILD=off
-```
[…]
