### 2022-01-31 mithun218/GiNaCDE#31 @mithun218
Thank you for your thoughtful suggestions. I have implemented your suggestions [here](https://github.com/mithun218/GiNaCDE/pull/36)
