### 2022-01-31 #3885 @mithun218
@carstenbauer  Thank you for your review and valuable suggestions.  I have addressed all the issues you referenced [here](https://github.com/openjournals/joss-reviews/issues/3885#issuecomment-1025749433) . Please check these, and if I have missed anything, please let me know.
