### 2022-02-04 #3885 @carstenbauer
@mithun218 Great, LGTM now.

@meg-simula My points have been adequately addressed. I therefore endorse GiNaCDE for publication in JOSS.
