### 2022-02-28 #3885 @meg-simula
@mithun218 Thank you for your submission and for successfully handling all reviewer comments.

I have read through the paper, and find a number of relatively minor grammatical errors. Could you please proof-read the manuscript carefully for grammar before I proceed with this submission?

Some specific points to address are
[ ] Replace "various researches" with "various research"
[ ] Replace "scientific community fields" with "scientific fields"
[ ] Insert "a" to replace "is dependent variable" by "is a dependent variable"
[ ] Replace "a polynomial about u" by "a polynomial in u"
[ ] Insert "and" before "Kudryashov-Sinelshchikov equation".
[ ] Replace "engineer" by "engineering"
