+++ JOSS-paper/paper.md
@@ -36,8 +36,9 @@ The F-expansion method was first proposed by Zhou et al. [@fexpn024]. The first 
 
 In order to solve the NLPDEs, many computer packages are available.
 In 1996, Parkes and Duffy [@atfm] had implemented tanh-expansion in their Mathematica package ATFM. Later complete implementation of tanh-expansion has been done by Li and Liu (2002) [@rath] designing the Maple package RATH. Baldwin et al. [@pdespclpkg1] have developed the Mathematica package *PDESpecialSolutions.m* which admits polynomial solutions in tanh, sech, combinations thereof, JacobiSN, JacobiCN. RAEEM [@raeem] is one of the most popular packages written in the Maple programming language, which is a comprehensive and complete implementation of some powerful methods such as the tanh-method, the extended tanh-method, the Jacobi elliptic function method, and the elliptic equation method. One can note that most packages have been developed within commercially available software frameworks, such as Maple and Mathematica. Besides this, all these computer packages have implemented the function-expansion methods. One serious drawback of the function-expansion method is that the solutions which contain functions other than some specific type of functions, such as tanh, sech, JacobiSN, JacobiCN, are not obtained. Additionally, the non-polynomial forms of these particular functions are not obtained also. On the other hand, we have observed that F-expansion, mF-expansion, and FIM are different kinds of methods, which can overcome the limitations of the function-expansion method. To the best of our knowledge, the computer packages implementing F-expansion and first integral methods are not available so far.
+Keeping in mind all the above points of view, we have been motivated to develop a free and open-source computer package or C++ library called GiNaCDE, which implements the F-expansion and first integral methods. 
 
-Keeping in mind all the above points of view, we have been motivated to develop a free and open-source computer package or C++ library called GiNaCDE, which implements the F-expansion and first integral methods. The symbolic manipulations of GiNaCDE depend only on GiNaC [@ginac]. There are several advantages to use GiNaC over other CAS. GiNaC is a free and open-source pure C++ library. It can accept C++ programming language, a general-purpose object-oriented programming (OOP) language, and it is fast like commercially available computer algebra systems.
+The symbolic manipulations of GiNaCDE depend only on GiNaC [@ginac]. There are several advantages to use GiNaC over other CAS. GiNaC is a free and open-source pure C++ library. It can accept C++ programming language, a general-purpose object-oriented programming (OOP) language, and it is fast like commercially available computer algebra systems.
 Besides the library version of GiNaCDE, we have also developed a GUI version of GiNaCDE called GiNaCDE GUI, which facilitates users to solve NLPDEs automatically without writing programming and compilation each time. This GUI version guides us in each step to obtain the output results.
 
 
