### 2022-03-04 #3885 @mithun218
Thank you so much @meg-simula. I have carefully read through the paper and fixed all the grammatical errors [here](https://github.com/mithun218/GiNaCDE/pull/37) and [here](https://github.com/mithun218/GiNaCDE/pull/38). Please check these.
