+++ doc/gtoolsTutorial.tex
@@ -24,7 +24,7 @@
 \end{center}
 \vspace{3cm}
 \begin{center}
-\Large{gtools (V1.0.0)}
+\Large{gtools (V1.5.0)}
 \end{center}
 \vspace{2cm} 
 \begin{center} 
