### 2022-04-10 #3885 @mithun218
@meg-simula  I have addressed all the points you mentioned [here.](https://github.com/openjournals/joss-reviews/issues/3885?msclkid=754d7248b89811ec85a8b80f379ac4b7#issuecomment-1064043823) Please review and if I have missed anything please let me know.
