+++ JOSS-paper/paper.md
@@ -7,7 +7,7 @@ tags:
   - F-expansion method  
   - First integral method 
 authors:
-  - name: Mithun Bairagi^[corresponding author.]
+  - name: Mithun Bairagi
     orcid: 0000-0002-9678-4625
     affiliation: 1 # (Multiple affiliations must be quoted)
 affiliations:
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on the high-performance F-expansion [@fexpn024], modified F-expansion (mF-expansion) [@modfexpn] and First Integral Methods (FIM) [@fim0]. It unifies and gathers various research on the F-expansion method and FIM [@fexpn024][@fexpn024_1][@fexpn123][@0246][@234][@modfexpn][@fim0][@mirza][@complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific fields by applying any one of the three available methods (F-expansion, mF-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
+GiNaCDE is a free and open-source C++ library that solves NLPDEs (Nonlinear Partial Differential Equations) and NLODEs (Nonlinear Ordinary Differential Equations) based on the high-performance F-expansion [@fexpn024], modified F-expansion (mF-expansion) [@modfexpn] and First Integral Methods (FIM) [@fim0]. It unifies and gathers various research on the F-expansion method and FIM [@fexpn024; @fexpn024_1; @fexpn123; @0246; @234; @modfexpn; @fim0; @mirza; @complexTwt1]. It implements two versions of the F-expansion method: F-expansion and modified F-expansion methods. It can be used to get exact traveling-wave solutions of a wide variety of NLPDEs arising in different scientific fields by applying any one of the three available methods (F-expansion, mF-expansion, and first integral method). GiNaCDE is built on a pure C++ symbolic library GiNaC [@ginac], and its algebraic manipulations are performed by the classes provided by the GiNaC library. It also has a rich Graphical User Interface (GUI) which makes it more user-friendly.
 
 The library has been designed to solve the NLPDEs, which have the following general form
 \begin{equation}\label{geneq}
@@ -32,7 +32,7 @@ nonlinearity terms, higher-derivative terms, and complex NLPDE. Some of them are
 # Statement of need 
 
 The NLODEs and NLPDEs play an important role in the theoretical sciences to explain many nonlinear phenomena in various fields of science, such as biology, chemistry, engineering, solid-state physics, plasma physics, optical fibers. The exact (closed-form) traveling-wave solutions of such NLPDEs give much extra information, which helps us to study the result more deeply. The knowledge of closed-form solutions of NLODEs and NLPDEs helps to test the degree of accuracy of numerical solvers and also facilitates stability analysis. In the past few decades, many powerful methods have been presented to seek exact solutions of NLPDEs, such as F-expansion method and first integral method. 
[…]
