### 2022-04-20 #3885 @arfon
@mithun218 – I've made some fixes to the citation formatting in the PR: https://github.com/mithun218/GiNaCDE/pull/40
