### 2022-04-21 #3885 @mithun218
> @mithun218 – I've made some fixes to the citation formatting in the PR:  mithun218/GiNaCDE#40

Thanks @arfon, the pull request has been merged.
